Add more APIs and examples
add has_text, has_image, has_rtf, has_text API.
Add listener example for html
