+++ Cargo.toml
@@ -17,6 +17,5 @@ clipboard-master = "3.1.3"
 base64 = "0.21.5"
 tempfile = "3.8.1"
 image = "0.24.7"
-arboard = "3.2.1"
 clipboard-files = "0.1.1"
 clipboard-rs ={ git = "https://github.com/ChurchTao/clipboard-rs" }
