+++ webview-dist/index.js
@@ -1,1 +1,1 @@
-function e(e,t,n,a){return new(n||(n=Promise))((function(r,s){function i(e){try{d(a.next(e))}catch(e){s(e)}}function o(e){try{d(a.throw(e))}catch(e){s(e)}}function d(e){var t;e.done?r(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(i,o)}d((a=a.apply(e,t||[])).next())}))}function t(e,t){var n,a,r,s,i={label:0,sent:function(){if(1&r[0])throw r[1];return r[1]},trys:[],ops:[]};return s={next:o(0),throw:o(1),return:o(2)},"function"==typeof Symbol&&(s[Symbol.iterator]=function(){return this}),s;function o(o){return function(d){return function(o){if(n)throw new TypeError("Generator is already executing.");for(;s&&(s=0,o[0]&&(i=0)),i;)try{if(n=1,a&&(r=2&o[0]?a.return:o[0]?a.throw||((r=a.return)&&r.call(a),0):a.next)&&!(r=r.call(a,o[1])).done)return r;switch(a=0,r&&(o=[2&o[0],r.value]),o[0]){case 0:case 1:r=o;break;case 4:return i.label++,{value:o[1],done:!1};case 5:i.label++,a=o[1],o=[0];continue;case 7:o=i.ops.pop(),i.trys.pop();continue;default:if(!(r=i.trys,(r=r.length>0&&r[r.length-1])||6!==o[0]&&2!==o[0])){i=0;continue}if(3===o[0]&&(!r||o[1]>r[0]&&o[1]<r[3])){i.label=o[1];break}if(6===o[0]&&i.label<r[1]){i.label=r[1],r=o;break}if(r&&i.label<r[2]){i.label=r[2],i.ops.push(o);break}r[2]&&i.ops.pop(),i.trys.pop();continue}o=t.call(e,i)}catch(e){o=[6,e],a=0}finally{n=r=0}if(5&o[0])throw o[1];return{value:o[0]?o[1]:void 0,done:!0}}([o,d])}}}var n,a;!function(e){e.assertEqual=e=>e,e.assertIs=function(e){},e.assertNever=function(e){throw new Error},e.arrayToEnum=e=>{const t={};for(const n of e)t[n]=n;return t},e.getValidEnumValues=t=>{const n=e.objectKeys(t).filter((e=>"number"!=typeof t[t[e]])),a={};for(const e of n)a[e]=t[e];return e.objectValues(a)},e.objectValues=t=>e.objectKeys(t).map((function(e){return t[e]})),e.objectKeys="function"==typeof Object.keys?e=>Object.keys(e):e=>{const t=[];for(const n in e)Object.prototype.hasOwnProperty.call(e,n)&&t.push(n);return t},e.find=(e,t)=>{for(const n of e)if(t(n))return n},e.isInteger="function"==typeof Number.isInteger?e=>Number.isInteger(e):e=>"number"==typeof e&&isFinite(e)&&Math.floor(e)===e,e.joinValues=function(e,t=" | "){return e.map((e=>"string"==typeof e?`'${e}'`:e)).join(t)},e.jsonStringifyReplacer=(e,t)=>"bigint"==typeof t?t.toString():t}(n||(n={})),function(e){e.mergeShapes=(e,t)=>({...e,...t})}(a||(a={}));const r=n.arrayToEnum(["string","nan","number","integer","float","boolean","date","bigint","symbol","function","undefined","null","array","object","unknown","promise","void","never","map","set"]),s=e=>{switch(typeof e){case"undefined":return r.undefined;case"string":return r.string;case"number":return isNaN(e)?r.nan:r.number;case"boolean":return r.boolean;case"function":return r.function;case"bigint":return r.bigint;case"symbol":return r.symbol;case"object":return Array.isArray(e)?r.array:null===e?r.null:e.then&&"function"==typeof e.then&&e.catch&&"function"==typeof e.catch?r.promise:"undefined"!=typeof Map&&e instanceof Map?r.map:"undefined"!=typeof Set&&e instanceof Set?r.set:"undefined"!=typeof Date&&e instanceof Date?r.date:r.object;default:return r.unknown}},i=n.arrayToEnum(["invalid_type","invalid_literal","custom","invalid_union","invalid_union_discriminator","invalid_enum_value","unrecognized_keys","invalid_arguments","invalid_return_type","invalid_date","invalid_string","too_small","too_big","invalid_intersection_types","not_multiple_of","not_finite"]);class o extends Error{constructor(e){super(),this.issues=[],this.addIssue=e=>{this.issues=[...this.issues,e]},this.addIssues=(e=[])=>{this.issues=[...this.issues,...e]};const t=new.target.prototype;Object.setPrototypeOf?Object.setPrototypeOf(this,t):this.__proto__=t,this.name="ZodError",this.issues=e}get errors(){return this.issues}format(e){const t=e||function(e){return e.message},n={_errors:[]},a=e=>{for(const r of e.issues)if("invalid_union"===r.code)r.unionErrors.map(a);else if("invalid_return_type"===r.code)a(r.returnTypeError);else if("invalid_arguments"===r.code)a(r.argumentsError);else if(0===r.path.length)n._errors.push(t(r));else{let e=n,a=0;for(;a<r.path.length;){const n=r.path[a];a===r.path.length-1?(e[n]=e[n]||{_errors:[]},e[n]._errors.push(t(r))):e[n]=e[n]||{_errors:[]},e=e[n],a++}}};return a(this),n}toString(){return this.message}get message(){return JSON.stringify(this.issues,n.jsonStringifyReplacer,2)}get isEmpty(){return 0===this.issues.length}flatten(e=(e=>e.message)){const t={},n=[];for(const a of this.issues)a.path.length>0?(t[a.path[0]]=t[a.path[0]]||[],t[a.path[0]].push(e(a))):n.push(e(a));return{formErrors:n,fieldErrors:t}}get formErrors(){return this.flatten()}}o.create=e=>new o(e);const d=(e,t)=>{let a;switch(e.code){case i.invalid_type:a=e.received===r.undefined?"Required":`Expected ${e.expected}, received ${e.received}`;break;case i.invalid_literal:a=`Invalid literal value, expected ${JSON.stringify(e.expected,n.jsonStringifyReplacer)}`;break;case i.unrecognized_keys:a=`Unrecognized key(s) in object: ${n.joinValues(e.keys,", ")}`;break;case i.invalid_union:a="Invalid input";break;case i.invalid_union_discriminator:a=`Invalid discriminator value. Expected ${n.joinValues(e.options)}`;break;case i.invalid_enum_value:a=`Invalid enum value. Expected ${n.joinValues(e.options)}, received '${e.received}'`;break;case i.invalid_arguments:a="Invalid function arguments";break;case i.invalid_return_type:a="Invalid function return type";break;case i.invalid_date:a="Invalid date";break;case i.invalid_string:"object"==typeof e.validation?"includes"in e.validation?(a=`Invalid input: must include "${e.validation.includes}"`,"number"==typeof e.validation.position&&(a=`${a} at one or more positions greater than or equal to ${e.validation.position}`)):"startsWith"in e.validation?a=`Invalid input: must start with "${e.validation.startsWith}"`:"endsWith"in e.validation?a=`Invalid input: must end with "${e.validation.endsWith}"`:n.assertNever(e.validation):a="regex"!==e.validation?`Invalid ${e.validation}`:"Invalid";break;case i.too_small:a="array"===e.type?`Array must contain ${e.exact?"exactly":e.inclusive?"at least":"more than"} ${e.minimum} element(s)`:"string"===e.type?`String must contain ${e.exact?"exactly":e.inclusive?"at least":"over"} ${e.minimum} character(s)`:"number"===e.type?`Number must be ${e.exact?"exactly equal to ":e.inclusive?"greater than or equal to ":"greater than "}${e.minimum}`:"date"===e.type?`Date must be ${e.exact?"exactly equal to ":e.inclusive?"greater than or equal to ":"greater than "}${new Date(Number(e.minimum))}`:"Invalid input";break;case i.too_big:a="array"===e.type?`Array must contain ${e.exact?"exactly":e.inclusive?"at most":"less than"} ${e.maximum} element(s)`:"string"===e.type?`String must contain ${e.exact?"exactly":e.inclusive?"at most":"under"} ${e.maximum} character(s)`:"number"===e.type?`Number must be ${e.exact?"exactly":e.inclusive?"less than or equal to":"less than"} ${e.maximum}`:"bigint"===e.type?`BigInt must be ${e.exact?"exactly":e.inclusive?"less than or equal to":"less than"} ${e.maximum}`:"date"===e.type?`Date must be ${e.exact?"exactly":e.inclusive?"smaller than or equal to":"smaller than"} ${new Date(Number(e.maximum))}`:"Invalid input";break;case i.custom:a="Invalid input";break;case i.invalid_intersection_types:a="Intersection results could not be merged";break;case i.not_multiple_of:a=`Number must be a multiple of ${e.multipleOf}`;break;case i.not_finite:a="Number must be finite";break;default:a=t.defaultError,n.assertNever(e)}return{message:a}};let u=d;function c(){return u}const l=e=>{const{data:t,path:n,errorMaps:a,issueData:r}=e,s=[...n,...r.path||[]],i={...r,path:s};let o="";const d=a.filter((e=>!!e)).slice().reverse();for(const e of d)o=e(i,{data:t,defaultError:o}).message;return{...r,path:s,message:r.message||o}};function p(e,t){const n=l({issueData:t,data:e.data,path:e.path,errorMaps:[e.common.contextualErrorMap,e.schemaErrorMap,c(),d].filter((e=>!!e))});e.common.issues.push(n)}class h{constructor(){this.value="valid"}dirty(){"valid"===this.value&&(this.value="dirty")}abort(){"aborted"!==this.value&&(this.value="aborted")}static mergeArray(e,t){const n=[];for(const a of t){if("aborted"===a.status)return m;"dirty"===a.status&&e.dirty(),n.push(a.value)}return{status:e.value,value:n}}static async mergeObjectAsync(e,t){const n=[];for(const e of t)n.push({key:await e.key,value:await e.value});return h.mergeObjectSync(e,n)}static mergeObjectSync(e,t){const n={};for(const a of t){const{key:t,value:r}=a;if("aborted"===t.status)return m;if("aborted"===r.status)return m;"dirty"===t.status&&e.dirty(),"dirty"===r.status&&e.dirty(),"__proto__"===t.value||void 0===r.value&&!a.alwaysSet||(n[t.value]=r.value)}return{status:e.value,value:n}}}const m=Object.freeze({status:"aborted"}),f=e=>({status:"dirty",value:e}),y=e=>({status:"valid",value:e}),_=e=>"aborted"===e.status,v=e=>"dirty"===e.status,g=e=>"valid"===e.status,b=e=>"undefined"!=typeof Promise&&e instanceof Promise;var x;!function(e){e.errToObj=e=>"string"==typeof e?{message:e}:e||{},e.toString=e=>"string"==typeof e?e:null==e?void 0:e.message}(x||(x={}));class k{constructor(e,t,n,a){this._cachedPath=[],this.parent=e,this.data=t,this._path=n,this._key=a}get path(){return this._cachedPath.length||(this._key instanceof Array?this._cachedPath.push(...this._path,...this._key):this._cachedPath.push(...this._path,this._key)),this._cachedPath}}const w=(e,t)=>{if(g(t))return{success:!0,data:t.value};if(!e.common.issues.length)throw new Error("Validation failed but no issues detected.");return{success:!1,get error(){if(this._error)return this._error;const t=new o(e.common.issues);return this._error=t,this._error}}};function Z(e){if(!e)return{};const{errorMap:t,invalid_type_error:n,required_error:a,description:r}=e;if(t&&(n||a))throw new Error('Can\'t use "invalid_type_error" or "required_error" in conjunction with custom error map.');if(t)return{errorMap:t,description:r};return{errorMap:(e,t)=>"invalid_type"!==e.code?{message:t.defaultError}:void 0===t.data?{message:null!=a?a:t.defaultError}:{message:null!=n?n:t.defaultError},description:r}}class T{constructor(e){this.spa=this.safeParseAsync,this._def=e,this.parse=this.parse.bind(this),this.safeParse=this.safeParse.bind(this),this.parseAsync=this.parseAsync.bind(this),this.safeParseAsync=this.safeParseAsync.bind(this),this.spa=this.spa.bind(this),this.refine=this.refine.bind(this),this.refinement=this.refinement.bind(this),this.superRefine=this.superRefine.bind(this),this.optional=this.optional.bind(this),this.nullable=this.nullable.bind(this),this.nullish=this.nullish.bind(this),this.array=this.array.bind(this),this.promise=this.promise.bind(this),this.or=this.or.bind(this),this.and=this.and.bind(this),this.transform=this.transform.bind(this),this.brand=this.brand.bind(this),this.default=this.default.bind(this),this.catch=this.catch.bind(this),this.describe=this.describe.bind(this),this.pipe=this.pipe.bind(this),this.readonly=this.readonly.bind(this),this.isNullable=this.isNullable.bind(this),this.isOptional=this.isOptional.bind(this)}get description(){return this._def.description}_getType(e){return s(e.data)}_getOrReturnCtx(e,t){return t||{common:e.parent.common,data:e.data,parsedType:s(e.data),schemaErrorMap:this._def.errorMap,path:e.path,parent:e.parent}}_processInputParams(e){return{status:new h,ctx:{common:e.parent.common,data:e.data,parsedType:s(e.data),schemaErrorMap:this._def.errorMap,path:e.path,parent:e.parent}}}_parseSync(e){const t=this._parse(e);if(b(t))throw new Error("Synchronous parse encountered promise.");return t}_parseAsync(e){const t=this._parse(e);return Promise.resolve(t)}parse(e,t){const n=this.safeParse(e,t);if(n.success)return n.data;throw n.error}safeParse(e,t){var n;const a={common:{issues:[],async:null!==(n=null==t?void 0:t.async)&&void 0!==n&&n,contextualErrorMap:null==t?void 0:t.errorMap},path:(null==t?void 0:t.path)||[],schemaErrorMap:this._def.errorMap,parent:null,data:e,parsedType:s(e)},r=this._parseSync({data:e,path:a.path,parent:a});return w(a,r)}async parseAsync(e,t){const n=await this.safeParseAsync(e,t);if(n.success)return n.data;throw n.error}async safeParseAsync(e,t){const n={common:{issues:[],contextualErrorMap:null==t?void 0:t.errorMap,async:!0},path:(null==t?void 0:t.path)||[],schemaErrorMap:this._def.errorMap,parent:null,data:e,parsedType:s(e)},a=this._parse({data:e,path:n.path,parent:n}),r=await(b(a)?a:Promise.resolve(a));return w(n,r)}refine(e,t){const n=e=>"string"==typeof t||void 0===t?{message:t}:"function"==typeof t?t(e):t;return this._refinement(((t,a)=>{const r=e(t),s=()=>a.addIssue({code:i.custom,...n(t)});return"undefined"!=typeof Promise&&r instanceof Promise?r.then((e=>!!e||(s(),!1))):!!r||(s(),!1)}))}refinement(e,t){return this._refinement(((n,a)=>!!e(n)||(a.addIssue("function"==typeof t?t(n,a):t),!1)))}_refinement(e){return new pe({schema:this,typeName:Ze.ZodEffects,effect:{type:"refinement",refinement:e}})}superRefine(e){return this._refinement(e)}optional(){return he.create(this,this._def)}nullable(){return me.create(this,this._def)}nullish(){return this.nullable().optional()}array(){return q.create(this,this._def)}promise(){return le.create(this,this._def)}or(e){return J.create([this,e],this._def)}and(e){return ee.create(this,e,this._def)}transform(e){return new pe({...Z(this._def),schema:this,typeName:Ze.ZodEffects,effect:{type:"transform",transform:e}})}default(e){const t="function"==typeof e?e:()=>e;return new fe({...Z(this._def),innerType:this,defaultValue:t,typeName:Ze.ZodDefault})}brand(){return new ge({typeName:Ze.ZodBranded,type:this,...Z(this._def)})}catch(e){const t="function"==typeof e?e:()=>e;return new ye({...Z(this._def),innerType:this,catchValue:t,typeName:Ze.ZodCatch})}describe(e){return new(0,this.constructor)({...this._def,description:e})}pipe(e){return be.create(this,e)}readonly(){return xe.create(this)}isOptional(){return this.safeParse(void 0).success}isNullable(){return this.safeParse(null).success}}const O=/^c[^\s-]{8,}$/i,E=/^[a-z][a-z0-9]*$/,N=/^[0-9A-HJKMNP-TV-Z]{26}$/,S=/^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$/i,C=/^(?!\.)(?!.*\.\.)([A-Z0-9_+-\.]*)[A-Z0-9_+-]@([A-Z0-9][A-Z0-9\-]*\.)+[A-Z]{2,}$/i;let I;const j=/^(((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2}))\.){3}((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2}))$/,R=/^(([a-f0-9]{1,4}:){7}|::([a-f0-9]{1,4}:){0,6}|([a-f0-9]{1,4}:){1}:([a-f0-9]{1,4}:){0,5}|([a-f0-9]{1,4}:){2}:([a-f0-9]{1,4}:){0,4}|([a-f0-9]{1,4}:){3}:([a-f0-9]{1,4}:){0,3}|([a-f0-9]{1,4}:){4}:([a-f0-9]{1,4}:){0,2}|([a-f0-9]{1,4}:){5}:([a-f0-9]{1,4}:){0,1})([a-f0-9]{1,4}|(((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2}))\.){3}((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2})))$/;class P extends T{_parse(e){this._def.coerce&&(e.data=String(e.data));if(this._getType(e)!==r.string){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.string,received:t.parsedType}),m}const t=new h;let a;for(const r of this._def.checks)if("min"===r.kind)e.data.length<r.value&&(a=this._getOrReturnCtx(e,a),p(a,{code:i.too_small,minimum:r.value,type:"string",inclusive:!0,exact:!1,message:r.message}),t.dirty());else if("max"===r.kind)e.data.length>r.value&&(a=this._getOrReturnCtx(e,a),p(a,{code:i.too_big,maximum:r.value,type:"string",inclusive:!0,exact:!1,message:r.message}),t.dirty());else if("length"===r.kind){const n=e.data.length>r.value,s=e.data.length<r.value;(n||s)&&(a=this._getOrReturnCtx(e,a),n?p(a,{code:i.too_big,maximum:r.value,type:"string",inclusive:!0,exact:!0,message:r.message}):s&&p(a,{code:i.too_small,minimum:r.value,type:"string",inclusive:!0,exact:!0,message:r.message}),t.dirty())}else if("email"===r.kind)C.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"email",code:i.invalid_string,message:r.message}),t.dirty());else if("emoji"===r.kind)I||(I=new RegExp("^(\\p{Extended_Pictographic}|\\p{Emoji_Component})+$","u")),I.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"emoji",code:i.invalid_string,message:r.message}),t.dirty());else if("uuid"===r.kind)S.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"uuid",code:i.invalid_string,message:r.message}),t.dirty());else if("cuid"===r.kind)O.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"cuid",code:i.invalid_string,message:r.message}),t.dirty());else if("cuid2"===r.kind)E.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"cuid2",code:i.invalid_string,message:r.message}),t.dirty());else if("ulid"===r.kind)N.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"ulid",code:i.invalid_string,message:r.message}),t.dirty());else if("url"===r.kind)try{new URL(e.data)}catch(n){a=this._getOrReturnCtx(e,a),p(a,{validation:"url",code:i.invalid_string,message:r.message}),t.dirty()}else if("regex"===r.kind){r.regex.lastIndex=0;r.regex.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"regex",code:i.invalid_string,message:r.message}),t.dirty())}else if("trim"===r.kind)e.data=e.data.trim();else if("includes"===r.kind)e.data.includes(r.value,r.position)||(a=this._getOrReturnCtx(e,a),p(a,{code:i.invalid_string,validation:{includes:r.value,position:r.position},message:r.message}),t.dirty());else if("toLowerCase"===r.kind)e.data=e.data.toLowerCase();else if("toUpperCase"===r.kind)e.data=e.data.toUpperCase();else if("startsWith"===r.kind)e.data.startsWith(r.value)||(a=this._getOrReturnCtx(e,a),p(a,{code:i.invalid_string,validation:{startsWith:r.value},message:r.message}),t.dirty());else if("endsWith"===r.kind)e.data.endsWith(r.value)||(a=this._getOrReturnCtx(e,a),p(a,{code:i.invalid_string,validation:{endsWith:r.value},message:r.message}),t.dirty());else if("datetime"===r.kind){((d=r).precision?d.offset?new RegExp(`^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{${d.precision}}(([+-]\\d{2}(:?\\d{2})?)|Z)$`):new RegExp(`^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{${d.precision}}Z$`):0===d.precision?d.offset?new RegExp("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(([+-]\\d{2}(:?\\d{2})?)|Z)$"):new RegExp("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"):d.offset?new RegExp("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d+)?(([+-]\\d{2}(:?\\d{2})?)|Z)$"):new RegExp("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d+)?Z$")).test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{code:i.invalid_string,validation:"datetime",message:r.message}),t.dirty())}else"ip"===r.kind?(s=e.data,("v4"!==(o=r.version)&&o||!j.test(s))&&("v6"!==o&&o||!R.test(s))&&(a=this._getOrReturnCtx(e,a),p(a,{validation:"ip",code:i.invalid_string,message:r.message}),t.dirty())):n.assertNever(r);var s,o,d;return{status:t.value,value:e.data}}_regex(e,t,n){return this.refinement((t=>e.test(t)),{validation:t,code:i.invalid_string,...x.errToObj(n)})}_addCheck(e){return new P({...this._def,checks:[...this._def.checks,e]})}email(e){return this._addCheck({kind:"email",...x.errToObj(e)})}url(e){return this._addCheck({kind:"url",...x.errToObj(e)})}emoji(e){return this._addCheck({kind:"emoji",...x.errToObj(e)})}uuid(e){return this._addCheck({kind:"uuid",...x.errToObj(e)})}cuid(e){return this._addCheck({kind:"cuid",...x.errToObj(e)})}cuid2(e){return this._addCheck({kind:"cuid2",...x.errToObj(e)})}ulid(e){return this._addCheck({kind:"ulid",...x.errToObj(e)})}ip(e){return this._addCheck({kind:"ip",...x.errToObj(e)})}datetime(e){var t;return"string"==typeof e?this._addCheck({kind:"datetime",precision:null,offset:!1,message:e}):this._addCheck({kind:"datetime",precision:void 0===(null==e?void 0:e.precision)?null:null==e?void 0:e.precision,offset:null!==(t=null==e?void 0:e.offset)&&void 0!==t&&t,...x.errToObj(null==e?void 0:e.message)})}regex(e,t){return this._addCheck({kind:"regex",regex:e,...x.errToObj(t)})}includes(e,t){return this._addCheck({kind:"includes",value:e,position:null==t?void 0:t.position,...x.errToObj(null==t?void 0:t.message)})}startsWith(e,t){return this._addCheck({kind:"startsWith",value:e,...x.errToObj(t)})}endsWith(e,t){return this._addCheck({kind:"endsWith",value:e,...x.errToObj(t)})}min(e,t){return this._addCheck({kind:"min",value:e,...x.errToObj(t)})}max(e,t){return this._addCheck({kind:"max",value:e,...x.errToObj(t)})}length(e,t){return this._addCheck({kind:"length",value:e,...x.errToObj(t)})}nonempty(e){return this.min(1,x.errToObj(e))}trim(){return new P({...this._def,checks:[...this._def.checks,{kind:"trim"}]})}toLowerCase(){return new P({...this._def,checks:[...this._def.checks,{kind:"toLowerCase"}]})}toUpperCase(){return new P({...this._def,checks:[...this._def.checks,{kind:"toUpperCase"}]})}get isDatetime(){return!!this._def.checks.find((e=>"datetime"===e.kind))}get isEmail(){return!!this._def.checks.find((e=>"email"===e.kind))}get isURL(){return!!this._def.checks.find((e=>"url"===e.kind))}get isEmoji(){return!!this._def.checks.find((e=>"emoji"===e.kind))}get isUUID(){return!!this._def.checks.find((e=>"uuid"===e.kind))}get isCUID(){return!!this._def.checks.find((e=>"cuid"===e.kind))}get isCUID2(){return!!this._def.checks.find((e=>"cuid2"===e.kind))}get isULID(){return!!this._def.checks.find((e=>"ulid"===e.kind))}get isIP(){return!!this._def.checks.find((e=>"ip"===e.kind))}get minLength(){let e=null;for(const t of this._def.checks)"min"===t.kind&&(null===e||t.value>e)&&(e=t.value);return e}get maxLength(){let e=null;for(const t of this._def.checks)"max"===t.kind&&(null===e||t.value<e)&&(e=t.value);return e}}function A(e,t){const n=(e.toString().split(".")[1]||"").length,a=(t.toString().split(".")[1]||"").length,r=n>a?n:a;return parseInt(e.toFixed(r).replace(".",""))%parseInt(t.toFixed(r).replace(".",""))/Math.pow(10,r)}P.create=e=>{var t;return new P({checks:[],typeName:Ze.ZodString,coerce:null!==(t=null==e?void 0:e.coerce)&&void 0!==t&&t,...Z(e)})};class D extends T{constructor(){super(...arguments),this.min=this.gte,this.max=this.lte,this.step=this.multipleOf}_parse(e){this._def.coerce&&(e.data=Number(e.data));if(this._getType(e)!==r.number){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.number,received:t.parsedType}),m}let t;const a=new h;for(const r of this._def.checks)if("int"===r.kind)n.isInteger(e.data)||(t=this._getOrReturnCtx(e,t),p(t,{code:i.invalid_type,expected:"integer",received:"float",message:r.message}),a.dirty());else if("min"===r.kind){(r.inclusive?e.data<r.value:e.data<=r.value)&&(t=this._getOrReturnCtx(e,t),p(t,{code:i.too_small,minimum:r.value,type:"number",inclusive:r.inclusive,exact:!1,message:r.message}),a.dirty())}else if("max"===r.kind){(r.inclusive?e.data>r.value:e.data>=r.value)&&(t=this._getOrReturnCtx(e,t),p(t,{code:i.too_big,maximum:r.value,type:"number",inclusive:r.inclusive,exact:!1,message:r.message}),a.dirty())}else"multipleOf"===r.kind?0!==A(e.data,r.value)&&(t=this._getOrReturnCtx(e,t),p(t,{code:i.not_multiple_of,multipleOf:r.value,message:r.message}),a.dirty()):"finite"===r.kind?Number.isFinite(e.data)||(t=this._getOrReturnCtx(e,t),p(t,{code:i.not_finite,message:r.message}),a.dirty()):n.assertNever(r);return{status:a.value,value:e.data}}gte(e,t){return this.setLimit("min",e,!0,x.toString(t))}gt(e,t){return this.setLimit("min",e,!1,x.toString(t))}lte(e,t){return this.setLimit("max",e,!0,x.toString(t))}lt(e,t){return this.setLimit("max",e,!1,x.toString(t))}setLimit(e,t,n,a){return new D({...this._def,checks:[...this._def.checks,{kind:e,value:t,inclusive:n,message:x.toString(a)}]})}_addCheck(e){return new D({...this._def,checks:[...this._def.checks,e]})}int(e){return this._addCheck({kind:"int",message:x.toString(e)})}positive(e){return this._addCheck({kind:"min",value:0,inclusive:!1,message:x.toString(e)})}negative(e){return this._addCheck({kind:"max",value:0,inclusive:!1,message:x.toString(e)})}nonpositive(e){return this._addCheck({kind:"max",value:0,inclusive:!0,message:x.toString(e)})}nonnegative(e){return this._addCheck({kind:"min",value:0,inclusive:!0,message:x.toString(e)})}multipleOf(e,t){return this._addCheck({kind:"multipleOf",value:e,message:x.toString(t)})}finite(e){return this._addCheck({kind:"finite",message:x.toString(e)})}safe(e){return this._addCheck({kind:"min",inclusive:!0,value:Number.MIN_SAFE_INTEGER,message:x.toString(e)})._addCheck({kind:"max",inclusive:!0,value:Number.MAX_SAFE_INTEGER,message:x.toString(e)})}get minValue(){let e=null;for(const t of this._def.checks)"min"===t.kind&&(null===e||t.value>e)&&(e=t.value);return e}get maxValue(){let e=null;for(const t of this._def.checks)"max"===t.kind&&(null===e||t.value<e)&&(e=t.value);return e}get isInt(){return!!this._def.checks.find((e=>"int"===e.kind||"multipleOf"===e.kind&&n.isInteger(e.value)))}get isFinite(){let e=null,t=null;for(const n of this._def.checks){if("finite"===n.kind||"int"===n.kind||"multipleOf"===n.kind)return!0;"min"===n.kind?(null===t||n.value>t)&&(t=n.value):"max"===n.kind&&(null===e||n.value<e)&&(e=n.value)}return Number.isFinite(t)&&Number.isFinite(e)}}D.create=e=>new D({checks:[],typeName:Ze.ZodNumber,coerce:(null==e?void 0:e.coerce)||!1,...Z(e)});class L extends T{constructor(){super(...arguments),this.min=this.gte,this.max=this.lte}_parse(e){this._def.coerce&&(e.data=BigInt(e.data));if(this._getType(e)!==r.bigint){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.bigint,received:t.parsedType}),m}let t;const a=new h;for(const r of this._def.checks)if("min"===r.kind){(r.inclusive?e.data<r.value:e.data<=r.value)&&(t=this._getOrReturnCtx(e,t),p(t,{code:i.too_small,type:"bigint",minimum:r.value,inclusive:r.inclusive,message:r.message}),a.dirty())}else if("max"===r.kind){(r.inclusive?e.data>r.value:e.data>=r.value)&&(t=this._getOrReturnCtx(e,t),p(t,{code:i.too_big,type:"bigint",maximum:r.value,inclusive:r.inclusive,message:r.message}),a.dirty())}else"multipleOf"===r.kind?e.data%r.value!==BigInt(0)&&(t=this._getOrReturnCtx(e,t),p(t,{code:i.not_multiple_of,multipleOf:r.value,message:r.message}),a.dirty()):n.assertNever(r);return{status:a.value,value:e.data}}gte(e,t){return this.setLimit("min",e,!0,x.toString(t))}gt(e,t){return this.setLimit("min",e,!1,x.toString(t))}lte(e,t){return this.setLimit("max",e,!0,x.toString(t))}lt(e,t){return this.setLimit("max",e,!1,x.toString(t))}setLimit(e,t,n,a){return new L({...this._def,checks:[...this._def.checks,{kind:e,value:t,inclusive:n,message:x.toString(a)}]})}_addCheck(e){return new L({...this._def,checks:[...this._def.checks,e]})}positive(e){return this._addCheck({kind:"min",value:BigInt(0),inclusive:!1,message:x.toString(e)})}negative(e){return this._addCheck({kind:"max",value:BigInt(0),inclusive:!1,message:x.toString(e)})}nonpositive(e){return this._addCheck({kind:"max",value:BigInt(0),inclusive:!0,message:x.toString(e)})}nonnegative(e){return this._addCheck({kind:"min",value:BigInt(0),inclusive:!0,message:x.toString(e)})}multipleOf(e,t){return this._addCheck({kind:"multipleOf",value:e,message:x.toString(t)})}get minValue(){let e=null;for(const t of this._def.checks)"min"===t.kind&&(null===e||t.value>e)&&(e=t.value);return e}get maxValue(){let e=null;for(const t of this._def.checks)"max"===t.kind&&(null===e||t.value<e)&&(e=t.value);return e}}L.create=e=>{var t;return new L({checks:[],typeName:Ze.ZodBigInt,coerce:null!==(t=null==e?void 0:e.coerce)&&void 0!==t&&t,...Z(e)})};class $ extends T{_parse(e){this._def.coerce&&(e.data=Boolean(e.data));if(this._getType(e)!==r.boolean){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.boolean,received:t.parsedType}),m}return y(e.data)}}$.create=e=>new $({typeName:Ze.ZodBoolean,coerce:(null==e?void 0:e.coerce)||!1,...Z(e)});class M extends T{_parse(e){this._def.coerce&&(e.data=new Date(e.data));if(this._getType(e)!==r.date){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.date,received:t.parsedType}),m}if(isNaN(e.data.getTime())){return p(this._getOrReturnCtx(e),{code:i.invalid_date}),m}const t=new h;let a;for(const r of this._def.checks)"min"===r.kind?e.data.getTime()<r.value&&(a=this._getOrReturnCtx(e,a),p(a,{code:i.too_small,message:r.message,inclusive:!0,exact:!1,minimum:r.value,type:"date"}),t.dirty()):"max"===r.kind?e.data.getTime()>r.value&&(a=this._getOrReturnCtx(e,a),p(a,{code:i.too_big,message:r.message,inclusive:!0,exact:!1,maximum:r.value,type:"date"}),t.dirty()):n.assertNever(r);return{status:t.value,value:new Date(e.data.getTime())}}_addCheck(e){return new M({...this._def,checks:[...this._def.checks,e]})}min(e,t){return this._addCheck({kind:"min",value:e.getTime(),message:x.toString(t)})}max(e,t){return this._addCheck({kind:"max",value:e.getTime(),message:x.toString(t)})}get minDate(){let e=null;for(const t of this._def.checks)"min"===t.kind&&(null===e||t.value>e)&&(e=t.value);return null!=e?new Date(e):null}get maxDate(){let e=null;for(const t of this._def.checks)"max"===t.kind&&(null===e||t.value<e)&&(e=t.value);return null!=e?new Date(e):null}}M.create=e=>new M({checks:[],coerce:(null==e?void 0:e.coerce)||!1,typeName:Ze.ZodDate,...Z(e)});class U extends T{_parse(e){if(this._getType(e)!==r.symbol){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.symbol,received:t.parsedType}),m}return y(e.data)}}U.create=e=>new U({typeName:Ze.ZodSymbol,...Z(e)});class W extends T{_parse(e){if(this._getType(e)!==r.undefined){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.undefined,received:t.parsedType}),m}return y(e.data)}}W.create=e=>new W({typeName:Ze.ZodUndefined,...Z(e)});class V extends T{_parse(e){if(this._getType(e)!==r.null){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.null,received:t.parsedType}),m}return y(e.data)}}V.create=e=>new V({typeName:Ze.ZodNull,...Z(e)});class z extends T{constructor(){super(...arguments),this._any=!0}_parse(e){return y(e.data)}}z.create=e=>new z({typeName:Ze.ZodAny,...Z(e)});class K extends T{constructor(){super(...arguments),this._unknown=!0}_parse(e){return y(e.data)}}K.create=e=>new K({typeName:Ze.ZodUnknown,...Z(e)});class B extends T{_parse(e){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.never,received:t.parsedType}),m}}B.create=e=>new B({typeName:Ze.ZodNever,...Z(e)});class F extends T{_parse(e){if(this._getType(e)!==r.undefined){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.void,received:t.parsedType}),m}return y(e.data)}}F.create=e=>new F({typeName:Ze.ZodVoid,...Z(e)});class q extends T{_parse(e){const{ctx:t,status:n}=this._processInputParams(e),a=this._def;if(t.parsedType!==r.array)return p(t,{code:i.invalid_type,expected:r.array,received:t.parsedType}),m;if(null!==a.exactLength){const e=t.data.length>a.exactLength.value,r=t.data.length<a.exactLength.value;(e||r)&&(p(t,{code:e?i.too_big:i.too_small,minimum:r?a.exactLength.value:void 0,maximum:e?a.exactLength.value:void 0,type:"array",inclusive:!0,exact:!0,message:a.exactLength.message}),n.dirty())}if(null!==a.minLength&&t.data.length<a.minLength.value&&(p(t,{code:i.too_small,minimum:a.minLength.value,type:"array",inclusive:!0,exact:!1,message:a.minLength.message}),n.dirty()),null!==a.maxLength&&t.data.length>a.maxLength.value&&(p(t,{code:i.too_big,maximum:a.maxLength.value,type:"array",inclusive:!0,exact:!1,message:a.maxLength.message}),n.dirty()),t.common.async)return Promise.all([...t.data].map(((e,n)=>a.type._parseAsync(new k(t,e,t.path,n))))).then((e=>h.mergeArray(n,e)));const s=[...t.data].map(((e,n)=>a.type._parseSync(new k(t,e,t.path,n))));return h.mergeArray(n,s)}get element(){return this._def.type}min(e,t){return new q({...this._def,minLength:{value:e,message:x.toString(t)}})}max(e,t){return new q({...this._def,maxLength:{value:e,message:x.toString(t)}})}length(e,t){return new q({...this._def,exactLength:{value:e,message:x.toString(t)}})}nonempty(e){return this.min(1,e)}}function H(e){if(e instanceof G){const t={};for(const n in e.shape){const a=e.shape[n];t[n]=he.create(H(a))}return new G({...e._def,shape:()=>t})}return e instanceof q?new q({...e._def,type:H(e.element)}):e instanceof he?he.create(H(e.unwrap())):e instanceof me?me.create(H(e.unwrap())):e instanceof te?te.create(e.items.map((e=>H(e)))):e}q.create=(e,t)=>new q({type:e,minLength:null,maxLength:null,exactLength:null,typeName:Ze.ZodArray,...Z(t)});class G extends T{constructor(){super(...arguments),this._cached=null,this.nonstrict=this.passthrough,this.augment=this.extend}_getCached(){if(null!==this._cached)return this._cached;const e=this._def.shape(),t=n.objectKeys(e);return this._cached={shape:e,keys:t}}_parse(e){if(this._getType(e)!==r.object){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.object,received:t.parsedType}),m}const{status:t,ctx:n}=this._processInputParams(e),{shape:a,keys:s}=this._getCached(),o=[];if(!(this._def.catchall instanceof B&&"strip"===this._def.unknownKeys))for(const e in n.data)s.includes(e)||o.push(e);const d=[];for(const e of s){const t=a[e],r=n.data[e];d.push({key:{status:"valid",value:e},value:t._parse(new k(n,r,n.path,e)),alwaysSet:e in n.data})}if(this._def.catchall instanceof B){const e=this._def.unknownKeys;if("passthrough"===e)for(const e of o)d.push({key:{status:"valid",value:e},value:{status:"valid",value:n.data[e]}});else if("strict"===e)o.length>0&&(p(n,{code:i.unrecognized_keys,keys:o}),t.dirty());else if("strip"!==e)throw new Error("Internal ZodObject error: invalid unknownKeys value.")}else{const e=this._def.catchall;for(const t of o){const a=n.data[t];d.push({key:{status:"valid",value:t},value:e._parse(new k(n,a,n.path,t)),alwaysSet:t in n.data})}}return n.common.async?Promise.resolve().then((async()=>{const e=[];for(const t of d){const n=await t.key;e.push({key:n,value:await t.value,alwaysSet:t.alwaysSet})}return e})).then((e=>h.mergeObjectSync(t,e))):h.mergeObjectSync(t,d)}get shape(){return this._def.shape()}strict(e){return x.errToObj,new G({...this._def,unknownKeys:"strict",...void 0!==e?{errorMap:(t,n)=>{var a,r,s,i;const o=null!==(s=null===(r=(a=this._def).errorMap)||void 0===r?void 0:r.call(a,t,n).message)&&void 0!==s?s:n.defaultError;return"unrecognized_keys"===t.code?{message:null!==(i=x.errToObj(e).message)&&void 0!==i?i:o}:{message:o}}}:{}})}strip(){return new G({...this._def,unknownKeys:"strip"})}passthrough(){return new G({...this._def,unknownKeys:"passthrough"})}extend(e){return new G({...this._def,shape:()=>({...this._def.shape(),...e})})}merge(e){return new G({unknownKeys:e._def.unknownKeys,catchall:e._def.catchall,shape:()=>({...this._def.shape(),...e._def.shape()}),typeName:Ze.ZodObject})}setKey(e,t){return this.augment({[e]:t})}catchall(e){return new G({...this._def,catchall:e})}pick(e){const t={};return n.objectKeys(e).forEach((n=>{e[n]&&this.shape[n]&&(t[n]=this.shape[n])})),new G({...this._def,shape:()=>t})}omit(e){const t={};return n.objectKeys(this.shape).forEach((n=>{e[n]||(t[n]=this.shape[n])})),new G({...this._def,shape:()=>t})}deepPartial(){return H(this)}partial(e){const t={};return n.objectKeys(this.shape).forEach((n=>{const a=this.shape[n];e&&!e[n]?t[n]=a:t[n]=a.optional()})),new G({...this._def,shape:()=>t})}required(e){const t={};return n.objectKeys(this.shape).forEach((n=>{if(e&&!e[n])t[n]=this.shape[n];else{let e=this.shape[n];for(;e instanceof he;)e=e._def.innerType;t[n]=e}})),new G({...this._def,shape:()=>t})}keyof(){return de(n.objectKeys(this.shape))}}G.create=(e,t)=>new G({shape:()=>e,unknownKeys:"strip",catchall:B.create(),typeName:Ze.ZodObject,...Z(t)}),G.strictCreate=(e,t)=>new G({shape:()=>e,unknownKeys:"strict",catchall:B.create(),typeName:Ze.ZodObject,...Z(t)}),G.lazycreate=(e,t)=>new G({shape:e,unknownKeys:"strip",catchall:B.create(),typeName:Ze.ZodObject,...Z(t)});class J extends T{_parse(e){const{ctx:t}=this._processInputParams(e),n=this._def.options;if(t.common.async)return Promise.all(n.map((async e=>{const n={...t,common:{...t.common,issues:[]},parent:null};return{result:await e._parseAsync({data:t.data,path:t.path,parent:n}),ctx:n}}))).then((function(e){for(const t of e)if("valid"===t.result.status)return t.result;for(const n of e)if("dirty"===n.result.status)return t.common.issues.push(...n.ctx.common.issues),n.result;const n=e.map((e=>new o(e.ctx.common.issues)));return p(t,{code:i.invalid_union,unionErrors:n}),m}));{let e;const a=[];for(const r of n){const n={...t,common:{...t.common,issues:[]},parent:null},s=r._parseSync({data:t.data,path:t.path,parent:n});if("valid"===s.status)return s;"dirty"!==s.status||e||(e={result:s,ctx:n}),n.common.issues.length&&a.push(n.common.issues)}if(e)return t.common.issues.push(...e.ctx.common.issues),e.result;const r=a.map((e=>new o(e)));return p(t,{code:i.invalid_union,unionErrors:r}),m}}get options(){return this._def.options}}J.create=(e,t)=>new J({options:e,typeName:Ze.ZodUnion,...Z(t)});const Y=e=>e instanceof ie?Y(e.schema):e instanceof pe?Y(e.innerType()):e instanceof oe?[e.value]:e instanceof ue?e.options:e instanceof ce?Object.keys(e.enum):e instanceof fe?Y(e._def.innerType):e instanceof W?[void 0]:e instanceof V?[null]:null;class Q extends T{_parse(e){const{ctx:t}=this._processInputParams(e);if(t.parsedType!==r.object)return p(t,{code:i.invalid_type,expected:r.object,received:t.parsedType}),m;const n=this.discriminator,a=t.data[n],s=this.optionsMap.get(a);return s?t.common.async?s._parseAsync({data:t.data,path:t.path,parent:t}):s._parseSync({data:t.data,path:t.path,parent:t}):(p(t,{code:i.invalid_union_discriminator,options:Array.from(this.optionsMap.keys()),path:[n]}),m)}get discriminator(){return this._def.discriminator}get options(){return this._def.options}get optionsMap(){return this._def.optionsMap}static create(e,t,n){const a=new Map;for(const n of t){const t=Y(n.shape[e]);if(!t)throw new Error(`A discriminator value for key \`${e}\` could not be extracted from all schema options`);for(const r of t){if(a.has(r))throw new Error(`Discriminator property ${String(e)} has duplicate value ${String(r)}`);a.set(r,n)}}return new Q({typeName:Ze.ZodDiscriminatedUnion,discriminator:e,options:t,optionsMap:a,...Z(n)})}}function X(e,t){const a=s(e),i=s(t);if(e===t)return{valid:!0,data:e};if(a===r.object&&i===r.object){const a=n.objectKeys(t),r=n.objectKeys(e).filter((e=>-1!==a.indexOf(e))),s={...e,...t};for(const n of r){const a=X(e[n],t[n]);if(!a.valid)return{valid:!1};s[n]=a.data}return{valid:!0,data:s}}if(a===r.array&&i===r.array){if(e.length!==t.length)return{valid:!1};const n=[];for(let a=0;a<e.length;a++){const r=X(e[a],t[a]);if(!r.valid)return{valid:!1};n.push(r.data)}return{valid:!0,data:n}}return a===r.date&&i===r.date&&+e==+t?{valid:!0,data:e}:{valid:!1}}class ee extends T{_parse(e){const{status:t,ctx:n}=this._processInputParams(e),a=(e,a)=>{if(_(e)||_(a))return m;const r=X(e.value,a.value);return r.valid?((v(e)||v(a))&&t.dirty(),{status:t.value,value:r.data}):(p(n,{code:i.invalid_intersection_types}),m)};return n.common.async?Promise.all([this._def.left._parseAsync({data:n.data,path:n.path,parent:n}),this._def.right._parseAsync({data:n.data,path:n.path,parent:n})]).then((([e,t])=>a(e,t))):a(this._def.left._parseSync({data:n.data,path:n.path,parent:n}),this._def.right._parseSync({data:n.data,path:n.path,parent:n}))}}ee.create=(e,t,n)=>new ee({left:e,right:t,typeName:Ze.ZodIntersection,...Z(n)});class te extends T{_parse(e){const{status:t,ctx:n}=this._processInputParams(e);if(n.parsedType!==r.array)return p(n,{code:i.invalid_type,expected:r.array,received:n.parsedType}),m;if(n.data.length<this._def.items.length)return p(n,{code:i.too_small,minimum:this._def.items.length,inclusive:!0,exact:!1,type:"array"}),m;!this._def.rest&&n.data.length>this._def.items.length&&(p(n,{code:i.too_big,maximum:this._def.items.length,inclusive:!0,exact:!1,type:"array"}),t.dirty());const a=[...n.data].map(((e,t)=>{const a=this._def.items[t]||this._def.rest;return a?a._parse(new k(n,e,n.path,t)):null})).filter((e=>!!e));return n.common.async?Promise.all(a).then((e=>h.mergeArray(t,e))):h.mergeArray(t,a)}get items(){return this._def.items}rest(e){return new te({...this._def,rest:e})}}te.create=(e,t)=>{if(!Array.isArray(e))throw new Error("You must pass an array of schemas to z.tuple([ ... ])");return new te({items:e,typeName:Ze.ZodTuple,rest:null,...Z(t)})};class ne extends T{get keySchema(){return this._def.keyType}get valueSchema(){return this._def.valueType}_parse(e){const{status:t,ctx:n}=this._processInputParams(e);if(n.parsedType!==r.object)return p(n,{code:i.invalid_type,expected:r.object,received:n.parsedType}),m;const a=[],s=this._def.keyType,o=this._def.valueType;for(const e in n.data)a.push({key:s._parse(new k(n,e,n.path,e)),value:o._parse(new k(n,n.data[e],n.path,e))});return n.common.async?h.mergeObjectAsync(t,a):h.mergeObjectSync(t,a)}get element(){return this._def.valueType}static create(e,t,n){return new ne(t instanceof T?{keyType:e,valueType:t,typeName:Ze.ZodRecord,...Z(n)}:{keyType:P.create(),valueType:e,typeName:Ze.ZodRecord,...Z(t)})}}class ae extends T{get keySchema(){return this._def.keyType}get valueSchema(){return this._def.valueType}_parse(e){const{status:t,ctx:n}=this._processInputParams(e);if(n.parsedType!==r.map)return p(n,{code:i.invalid_type,expected:r.map,received:n.parsedType}),m;const a=this._def.keyType,s=this._def.valueType,o=[...n.data.entries()].map((([e,t],r)=>({key:a._parse(new k(n,e,n.path,[r,"key"])),value:s._parse(new k(n,t,n.path,[r,"value"]))})));if(n.common.async){const e=new Map;return Promise.resolve().then((async()=>{for(const n of o){const a=await n.key,r=await n.value;if("aborted"===a.status||"aborted"===r.status)return m;"dirty"!==a.status&&"dirty"!==r.status||t.dirty(),e.set(a.value,r.value)}return{status:t.value,value:e}}))}{const e=new Map;for(const n of o){const a=n.key,r=n.value;if("aborted"===a.status||"aborted"===r.status)return m;"dirty"!==a.status&&"dirty"!==r.status||t.dirty(),e.set(a.value,r.value)}return{status:t.value,value:e}}}}ae.create=(e,t,n)=>new ae({valueType:t,keyType:e,typeName:Ze.ZodMap,...Z(n)});class re extends T{_parse(e){const{status:t,ctx:n}=this._processInputParams(e);if(n.parsedType!==r.set)return p(n,{code:i.invalid_type,expected:r.set,received:n.parsedType}),m;const a=this._def;null!==a.minSize&&n.data.size<a.minSize.value&&(p(n,{code:i.too_small,minimum:a.minSize.value,type:"set",inclusive:!0,exact:!1,message:a.minSize.message}),t.dirty()),null!==a.maxSize&&n.data.size>a.maxSize.value&&(p(n,{code:i.too_big,maximum:a.maxSize.value,type:"set",inclusive:!0,exact:!1,message:a.maxSize.message}),t.dirty());const s=this._def.valueType;function o(e){const n=new Set;for(const a of e){if("aborted"===a.status)return m;"dirty"===a.status&&t.dirty(),n.add(a.value)}return{status:t.value,value:n}}const d=[...n.data.values()].map(((e,t)=>s._parse(new k(n,e,n.path,t))));return n.common.async?Promise.all(d).then((e=>o(e))):o(d)}min(e,t){return new re({...this._def,minSize:{value:e,message:x.toString(t)}})}max(e,t){return new re({...this._def,maxSize:{value:e,message:x.toString(t)}})}size(e,t){return this.min(e,t).max(e,t)}nonempty(e){return this.min(1,e)}}re.create=(e,t)=>new re({valueType:e,minSize:null,maxSize:null,typeName:Ze.ZodSet,...Z(t)});class se extends T{constructor(){super(...arguments),this.validate=this.implement}_parse(e){const{ctx:t}=this._processInputParams(e);if(t.parsedType!==r.function)return p(t,{code:i.invalid_type,expected:r.function,received:t.parsedType}),m;function n(e,n){return l({data:e,path:t.path,errorMaps:[t.common.contextualErrorMap,t.schemaErrorMap,c(),d].filter((e=>!!e)),issueData:{code:i.invalid_arguments,argumentsError:n}})}function a(e,n){return l({data:e,path:t.path,errorMaps:[t.common.contextualErrorMap,t.schemaErrorMap,c(),d].filter((e=>!!e)),issueData:{code:i.invalid_return_type,returnTypeError:n}})}const s={errorMap:t.common.contextualErrorMap},u=t.data;if(this._def.returns instanceof le){const e=this;return y((async function(...t){const r=new o([]),i=await e._def.args.parseAsync(t,s).catch((e=>{throw r.addIssue(n(t,e)),r})),d=await Reflect.apply(u,this,i),c=await e._def.returns._def.type.parseAsync(d,s).catch((e=>{throw r.addIssue(a(d,e)),r}));return c}))}{const e=this;return y((function(...t){const r=e._def.args.safeParse(t,s);if(!r.success)throw new o([n(t,r.error)]);const i=Reflect.apply(u,this,r.data),d=e._def.returns.safeParse(i,s);if(!d.success)throw new o([a(i,d.error)]);return d.data}))}}parameters(){return this._def.args}returnType(){return this._def.returns}args(...e){return new se({...this._def,args:te.create(e).rest(K.create())})}returns(e){return new se({...this._def,returns:e})}implement(e){return this.parse(e)}strictImplement(e){return this.parse(e)}static create(e,t,n){return new se({args:e||te.create([]).rest(K.create()),returns:t||K.create(),typeName:Ze.ZodFunction,...Z(n)})}}class ie extends T{get schema(){return this._def.getter()}_parse(e){const{ctx:t}=this._processInputParams(e);return this._def.getter()._parse({data:t.data,path:t.path,parent:t})}}ie.create=(e,t)=>new ie({getter:e,typeName:Ze.ZodLazy,...Z(t)});class oe extends T{_parse(e){if(e.data!==this._def.value){const t=this._getOrReturnCtx(e);return p(t,{received:t.data,code:i.invalid_literal,expected:this._def.value}),m}return{status:"valid",value:e.data}}get value(){return this._def.value}}function de(e,t){return new ue({values:e,typeName:Ze.ZodEnum,...Z(t)})}oe.create=(e,t)=>new oe({value:e,typeName:Ze.ZodLiteral,...Z(t)});class ue extends T{_parse(e){if("string"!=typeof e.data){const t=this._getOrReturnCtx(e),a=this._def.values;return p(t,{expected:n.joinValues(a),received:t.parsedType,code:i.invalid_type}),m}if(-1===this._def.values.indexOf(e.data)){const t=this._getOrReturnCtx(e),n=this._def.values;return p(t,{received:t.data,code:i.invalid_enum_value,options:n}),m}return y(e.data)}get options(){return this._def.values}get enum(){const e={};for(const t of this._def.values)e[t]=t;return e}get Values(){const e={};for(const t of this._def.values)e[t]=t;return e}get Enum(){const e={};for(const t of this._def.values)e[t]=t;return e}extract(e){return ue.create(e)}exclude(e){return ue.create(this.options.filter((t=>!e.includes(t))))}}ue.create=de;class ce extends T{_parse(e){const t=n.getValidEnumValues(this._def.values),a=this._getOrReturnCtx(e);if(a.parsedType!==r.string&&a.parsedType!==r.number){const e=n.objectValues(t);return p(a,{expected:n.joinValues(e),received:a.parsedType,code:i.invalid_type}),m}if(-1===t.indexOf(e.data)){const e=n.objectValues(t);return p(a,{received:a.data,code:i.invalid_enum_value,options:e}),m}return y(e.data)}get enum(){return this._def.values}}ce.create=(e,t)=>new ce({values:e,typeName:Ze.ZodNativeEnum,...Z(t)});class le extends T{unwrap(){return this._def.type}_parse(e){const{ctx:t}=this._processInputParams(e);if(t.parsedType!==r.promise&&!1===t.common.async)return p(t,{code:i.invalid_type,expected:r.promise,received:t.parsedType}),m;const n=t.parsedType===r.promise?t.data:Promise.resolve(t.data);return y(n.then((e=>this._def.type.parseAsync(e,{path:t.path,errorMap:t.common.contextualErrorMap}))))}}le.create=(e,t)=>new le({type:e,typeName:Ze.ZodPromise,...Z(t)});class pe extends T{innerType(){return this._def.schema}sourceType(){return this._def.schema._def.typeName===Ze.ZodEffects?this._def.schema.sourceType():this._def.schema}_parse(e){const{status:t,ctx:a}=this._processInputParams(e),r=this._def.effect||null,s={addIssue:e=>{p(a,e),e.fatal?t.abort():t.dirty()},get path(){return a.path}};if(s.addIssue=s.addIssue.bind(s),"preprocess"===r.type){const e=r.transform(a.data,s);return a.common.issues.length?{status:"dirty",value:a.data}:a.common.async?Promise.resolve(e).then((e=>this._def.schema._parseAsync({data:e,path:a.path,parent:a}))):this._def.schema._parseSync({data:e,path:a.path,parent:a})}if("refinement"===r.type){const e=e=>{const t=r.refinement(e,s);if(a.common.async)return Promise.resolve(t);if(t instanceof Promise)throw new Error("Async refinement encountered during synchronous parse operation. Use .parseAsync instead.");return e};if(!1===a.common.async){const n=this._def.schema._parseSync({data:a.data,path:a.path,parent:a});return"aborted"===n.status?m:("dirty"===n.status&&t.dirty(),e(n.value),{status:t.value,value:n.value})}return this._def.schema._parseAsync({data:a.data,path:a.path,parent:a}).then((n=>"aborted"===n.status?m:("dirty"===n.status&&t.dirty(),e(n.value).then((()=>({status:t.value,value:n.value}))))))}if("transform"===r.type){if(!1===a.common.async){const e=this._def.schema._parseSync({data:a.data,path:a.path,parent:a});if(!g(e))return e;const n=r.transform(e.value,s);if(n instanceof Promise)throw new Error("Asynchronous transform encountered during synchronous parse operation. Use .parseAsync instead.");return{status:t.value,value:n}}return this._def.schema._parseAsync({data:a.data,path:a.path,parent:a}).then((e=>g(e)?Promise.resolve(r.transform(e.value,s)).then((e=>({status:t.value,value:e}))):e))}n.assertNever(r)}}pe.create=(e,t,n)=>new pe({schema:e,typeName:Ze.ZodEffects,effect:t,...Z(n)}),pe.createWithPreprocess=(e,t,n)=>new pe({schema:t,effect:{type:"preprocess",transform:e},typeName:Ze.ZodEffects,...Z(n)});class he extends T{_parse(e){return this._getType(e)===r.undefined?y(void 0):this._def.innerType._parse(e)}unwrap(){return this._def.innerType}}he.create=(e,t)=>new he({innerType:e,typeName:Ze.ZodOptional,...Z(t)});class me extends T{_parse(e){return this._getType(e)===r.null?y(null):this._def.innerType._parse(e)}unwrap(){return this._def.innerType}}me.create=(e,t)=>new me({innerType:e,typeName:Ze.ZodNullable,...Z(t)});class fe extends T{_parse(e){const{ctx:t}=this._processInputParams(e);let n=t.data;return t.parsedType===r.undefined&&(n=this._def.defaultValue()),this._def.innerType._parse({data:n,path:t.path,parent:t})}removeDefault(){return this._def.innerType}}fe.create=(e,t)=>new fe({innerType:e,typeName:Ze.ZodDefault,defaultValue:"function"==typeof t.default?t.default:()=>t.default,...Z(t)});class ye extends T{_parse(e){const{ctx:t}=this._processInputParams(e),n={...t,common:{...t.common,issues:[]}},a=this._def.innerType._parse({data:n.data,path:n.path,parent:{...n}});return b(a)?a.then((e=>({status:"valid",value:"valid"===e.status?e.value:this._def.catchValue({get error(){return new o(n.common.issues)},input:n.data})}))):{status:"valid",value:"valid"===a.status?a.value:this._def.catchValue({get error(){return new o(n.common.issues)},input:n.data})}}removeCatch(){return this._def.innerType}}ye.create=(e,t)=>new ye({innerType:e,typeName:Ze.ZodCatch,catchValue:"function"==typeof t.catch?t.catch:()=>t.catch,...Z(t)});class _e extends T{_parse(e){if(this._getType(e)!==r.nan){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.nan,received:t.parsedType}),m}return{status:"valid",value:e.data}}}_e.create=e=>new _e({typeName:Ze.ZodNaN,...Z(e)});const ve=Symbol("zod_brand");class ge extends T{_parse(e){const{ctx:t}=this._processInputParams(e),n=t.data;return this._def.type._parse({data:n,path:t.path,parent:t})}unwrap(){return this._def.type}}class be extends T{_parse(e){const{status:t,ctx:n}=this._processInputParams(e);if(n.common.async){return(async()=>{const e=await this._def.in._parseAsync({data:n.data,path:n.path,parent:n});return"aborted"===e.status?m:"dirty"===e.status?(t.dirty(),f(e.value)):this._def.out._parseAsync({data:e.value,path:n.path,parent:n})})()}{const e=this._def.in._parseSync({data:n.data,path:n.path,parent:n});return"aborted"===e.status?m:"dirty"===e.status?(t.dirty(),{status:"dirty",value:e.value}):this._def.out._parseSync({data:e.value,path:n.path,parent:n})}}static create(e,t){return new be({in:e,out:t,typeName:Ze.ZodPipeline})}}class xe extends T{_parse(e){const t=this._def.innerType._parse(e);return g(t)&&(t.value=Object.freeze(t.value)),t}}xe.create=(e,t)=>new xe({innerType:e,typeName:Ze.ZodReadonly,...Z(t)});const ke=(e,t={},n)=>e?z.create().superRefine(((a,r)=>{var s,i;if(!e(a)){const e="function"==typeof t?t(a):"string"==typeof t?{message:t}:t,o=null===(i=null!==(s=e.fatal)&&void 0!==s?s:n)||void 0===i||i,d="string"==typeof e?{message:e}:e;r.addIssue({code:"custom",...d,fatal:o})}})):z.create(),we={object:G.lazycreate};var Ze;!function(e){e.ZodString="ZodString",e.ZodNumber="ZodNumber",e.ZodNaN="ZodNaN",e.ZodBigInt="ZodBigInt",e.ZodBoolean="ZodBoolean",e.ZodDate="ZodDate",e.ZodSymbol="ZodSymbol",e.ZodUndefined="ZodUndefined",e.ZodNull="ZodNull",e.ZodAny="ZodAny",e.ZodUnknown="ZodUnknown",e.ZodNever="ZodNever",e.ZodVoid="ZodVoid",e.ZodArray="ZodArray",e.ZodObject="ZodObject",e.ZodUnion="ZodUnion",e.ZodDiscriminatedUnion="ZodDiscriminatedUnion",e.ZodIntersection="ZodIntersection",e.ZodTuple="ZodTuple",e.ZodRecord="ZodRecord",e.ZodMap="ZodMap",e.ZodSet="ZodSet",e.ZodFunction="ZodFunction",e.ZodLazy="ZodLazy",e.ZodLiteral="ZodLiteral",e.ZodEnum="ZodEnum",e.ZodEffects="ZodEffects",e.ZodNativeEnum="ZodNativeEnum",e.ZodOptional="ZodOptional",e.ZodNullable="ZodNullable",e.ZodDefault="ZodDefault",e.ZodCatch="ZodCatch",e.ZodPromise="ZodPromise",e.ZodBranded="ZodBranded",e.ZodPipeline="ZodPipeline",e.ZodReadonly="ZodReadonly"}(Ze||(Ze={}));const Te=P.create,Oe=D.create,Ee=_e.create,Ne=L.create,Se=$.create,Ce=M.create,Ie=U.create,je=W.create,Re=V.create,Pe=z.create,Ae=K.create,De=B.create,Le=F.create,$e=q.create,Me=G.create,Ue=G.strictCreate,We=J.create,Ve=Q.create,ze=ee.create,Ke=te.create,Be=ne.create,Fe=ae.create,qe=re.create,He=se.create,Ge=ie.create,Je=oe.create,Ye=ue.create,Qe=ce.create,Xe=le.create,et=pe.create,tt=he.create,nt=me.create,at=pe.createWithPreprocess,rt=be.create,st={string:e=>P.create({...e,coerce:!0}),number:e=>D.create({...e,coerce:!0}),boolean:e=>$.create({...e,coerce:!0}),bigint:e=>L.create({...e,coerce:!0}),date:e=>M.create({...e,coerce:!0})},it=m;var ot=Object.freeze({__proto__:null,defaultErrorMap:d,setErrorMap:function(e){u=e},getErrorMap:c,makeIssue:l,EMPTY_PATH:[],addIssueToContext:p,ParseStatus:h,INVALID:m,DIRTY:f,OK:y,isAborted:_,isDirty:v,isValid:g,isAsync:b,get util(){return n},get objectUtil(){return a},ZodParsedType:r,getParsedType:s,ZodType:T,ZodString:P,ZodNumber:D,ZodBigInt:L,ZodBoolean:$,ZodDate:M,ZodSymbol:U,ZodUndefined:W,ZodNull:V,ZodAny:z,ZodUnknown:K,ZodNever:B,ZodVoid:F,ZodArray:q,ZodObject:G,ZodUnion:J,ZodDiscriminatedUnion:Q,ZodIntersection:ee,ZodTuple:te,ZodRecord:ne,ZodMap:ae,ZodSet:re,ZodFunction:se,ZodLazy:ie,ZodLiteral:oe,ZodEnum:ue,ZodNativeEnum:ce,ZodPromise:le,ZodEffects:pe,ZodTransformer:pe,ZodOptional:he,ZodNullable:me,ZodDefault:fe,ZodCatch:ye,ZodNaN:_e,BRAND:ve,ZodBranded:ge,ZodPipeline:be,ZodReadonly:xe,custom:ke,Schema:T,ZodSchema:T,late:we,get ZodFirstPartyTypeKind(){return Ze},coerce:st,any:Pe,array:$e,bigint:Ne,boolean:Se,date:Ce,discriminatedUnion:Ve,effect:et,enum:Ye,function:He,instanceof:(e,t={message:`Input not instance of ${e.name}`})=>ke((t=>t instanceof e),t),intersection:ze,lazy:Ge,literal:Je,map:Fe,nan:Ee,nativeEnum:Qe,never:De,null:Re,nullable:nt,number:Oe,object:Me,oboolean:()=>Se().optional(),onumber:()=>Oe().optional(),optional:tt,ostring:()=>Te().optional(),pipeline:rt,preprocess:at,promise:Xe,record:Be,set:qe,strictObject:Ue,string:Te,symbol:Ie,transformer:et,tuple:Ke,undefined:je,union:We,unknown:Ae,void:Le,NEVER:it,ZodIssueCode:i,quotelessJson:e=>JSON.stringify(e,null,2).replace(/"([^"]+)":/g,"$1:"),ZodError:o}),dt=Object.defineProperty,ut=(e,t)=>{for(var n in t)dt(e,n,{get:t[n],enumerable:!0})};function ct(e,t=!1){let n=window.crypto.getRandomValues(new Uint32Array(1))[0],a=`_${n}`;return Object.defineProperty(window,a,{value:n=>(t&&Reflect.deleteProperty(window,a),null==e?void 0:e(n)),writable:!1,configurable:!0}),n}async function lt(e,t={}){return new Promise(((n,a)=>{let r=ct((e=>{n(e),Reflect.deleteProperty(window,`_${s}`)}),!0),s=ct((e=>{a(e),Reflect.deleteProperty(window,`_${r}`)}),!0);window.__TAURI_IPC__({cmd:e,callback:r,error:s,...t})}))}function pt(e,t="asset"){let n=encodeURIComponent(e);return navigator.userAgent.includes("Windows")?`https://${t}.localhost/${n}`:`${t}://localhost/${n}`}async function ht(e){return lt("tauri",e)}ut({},{convertFileSrc:()=>pt,invoke:()=>lt,transformCallback:()=>ct});async function mt(e,t){return ht({__tauriModule:"Event",message:{cmd:"unlisten",event:e,eventId:t}})}async function ft(e,t,n){return ht({__tauriModule:"Event",message:{cmd:"listen",event:e,windowLabel:t,handler:ct(n)}}).then((t=>async()=>mt(e,t)))}ut({},{TauriEvent:()=>yt,emit:()=>gt,listen:()=>_t,once:()=>vt});var yt=(e=>(e.WINDOW_RESIZED="tauri://resize",e.WINDOW_MOVED="tauri://move",e.WINDOW_CLOSE_REQUESTED="tauri://close-requested",e.WINDOW_CREATED="tauri://window-created",e.WINDOW_DESTROYED="tauri://destroyed",e.WINDOW_FOCUS="tauri://focus",e.WINDOW_BLUR="tauri://blur",e.WINDOW_SCALE_FACTOR_CHANGED="tauri://scale-change",e.WINDOW_THEME_CHANGED="tauri://theme-changed",e.WINDOW_FILE_DROP="tauri://file-drop",e.WINDOW_FILE_DROP_HOVER="tauri://file-drop-hover",e.WINDOW_FILE_DROP_CANCELLED="tauri://file-drop-cancelled",e.MENU="tauri://menu",e.CHECK_UPDATE="tauri://update",e.UPDATE_AVAILABLE="tauri://update-available",e.INSTALL_UPDATE="tauri://update-install",e.STATUS_UPDATE="tauri://update-status",e.DOWNLOAD_PROGRESS="tauri://update-download-progress",e))(yt||{});async function _t(e,t){return ft(e,null,t)}async function vt(e,t){return async function(e,t,n){return ft(e,t,(t=>{n(t),mt(e,t.id).catch((()=>{}))}))}(e,null,t)}async function gt(e,t){return async function(e,t,n){await ht({__tauriModule:"Event",message:{cmd:"emit",event:e,windowLabel:t,payload:n}})}(e,void 0,t)}var bt="plugin:clipboard|start_monitor",xt="plugin:clipboard|stop_monitor",kt="plugin:clipboard://text-changed",wt="plugin:clipboard://files-changed",Zt="plugin:clipboard://image-changed",Tt="plugin:clipboard|is_monitor_running",Ot="plugin:clipboard|read_image_binary",Et="plugin:clipboard|write_text",Nt="plugin:clipboard|clear",St="plugin:clipboard|read_text",Ct="plugin:clipboard|read_files",It="plugin:clipboard|read_image",jt="plugin:clipboard|write_image",Rt="plugin:clipboard://clipboard-monitor/status",Pt="plugin:clipboard://clipboard-monitor/update",At=ot.object({value:ot.string()}),Dt=ot.object({value:ot.string().array()});function Lt(e){return lt(Et,{text:e})}function $t(){return lt(Nt)}function Mt(){return lt(St)}function Ut(){return lt(Ct)}function Wt(){return lt(It)}var Vt=Wt;function zt(e){return lt(Ot).then((function(t){switch(e){case"int_array":default:return t;case"Uint8Array":return new Uint8Array(t);case"Blob":return new Blob([new Uint8Array(t)])}}))}function Kt(){return zt("Blob").then((function(e){return URL.createObjectURL(e)}))}function Bt(e){return lt(jt,{base64Image:e})}function Ft(n){void 0===n&&(n=500);var a="",r=!0;return setTimeout((function s(){return e(this,void 0,void 0,(function(){var e;return t(this,(function(t){switch(t.label){case 0:return t.trys.push([0,4,,5]),[4,Mt()];case 1:return e=t.sent(),a===e?[3,3]:[4,gt(kt,{value:e})];case 2:t.sent(),t.label=3;case 3:return a=e,[3,5];case 4:return t.sent(),[3,5];case 5:return r&&setTimeout(s,n),[2]}}))}))}),n),function(){r=!1}}function qt(n){void 0===n&&(n=1e3);var a="",r=!0;return setTimeout((function s(){return e(this,void 0,void 0,(function(){var e;return t(this,(function(t){switch(t.label){case 0:return t.trys.push([0,4,,5]),[4,Wt()];case 1:return e=t.sent(),a===e?[3,3]:[4,gt(Zt,{value:e})];case 2:t.sent(),t.label=3;case 3:return a=e,[3,5];case 4:return t.sent(),[3,5];case 5:return r&&setTimeout(s,n),[2]}}))}))}),n),function(){r=!1}}function Ht(){var n=this;return _t(Pt,(function(a){return e(n,void 0,void 0,(function(){var e,n,r,s;return t(this,(function(t){switch(t.label){case 0:if("clipboard update"!==a.payload)return[3,17];t.label=1;case 1:return t.trys.push([1,4,,17]),[4,Ut()];case 2:return e=t.sent(),[4,gt(wt,{value:e})];case 3:return t.sent(),[3,17];case 4:t.sent(),t.label=5;case 5:return t.trys.push([5,9,,16]),[4,Mt()];case 6:return(n=t.sent())?[4,gt(kt,{value:n})]:[3,8];case 7:t.sent(),t.label=8;case 8:return[3,16];case 9:t.sent(),t.label=10;case 10:return t.trys.push([10,14,,15]),[4,Wt()];case 11:return(r=t.sent())?[4,gt(Zt,{value:r})]:[3,13];case 12:t.sent(),t.label=13;case 13:return[3,15];case 14:return s=t.sent(),console.error(s),[3,15];case 15:return[3,16];case 16:return[3,17];case 17:return[2]}}))}))}))}function Gt(e){return _t(Pt,e)}function Jt(n){return e(this,void 0,void 0,(function(){return t(this,(function(e){switch(e.label){case 0:return[4,_t(kt,(function(e){var t=At.parse(e.payload).value;n(t)}))];case 1:return[2,e.sent()]}}))}))}function Yt(n){return e(this,void 0,void 0,(function(){return t(this,(function(e){switch(e.label){case 0:return[4,_t(wt,(function(e){var t=Dt.parse(e.payload).value;n(t)}))];case 1:return[2,e.sent()]}}))}))}function Qt(n){return e(this,void 0,void 0,(function(){return t(this,(function(e){switch(e.label){case 0:return[4,_t(Zt,(function(e){var t=At.parse(e.payload).value;n(t)}))];case 1:return[2,e.sent()]}}))}))}function Xt(){return lt(Tt).then((function(e){return ot.boolean().parse(e)}))}function en(){return lt(bt)}function tn(){return lt(xt)}function nn(n){return e(this,void 0,void 0,(function(){return t(this,(function(e){switch(e.label){case 0:return[4,_t(Rt,(function(e){var t=ot.boolean().parse(e.payload);n(t)}))];case 1:return[2,e.sent()]}}))}))}function an(){var n=this;return en().then((function(){return Ht()})).then((function(a){return function(){return e(n,void 0,void 0,(function(){return t(this,(function(e){switch(e.label){case 0:return console.log("stop listening"),a(),[4,tn()];case 1:return e.sent(),[2]}}))}))}}))}export{Nt as CLEAR_COMMAND,Rt as CLIPBOARD_MONITOR_STATUS_UPDATE_EVENT,Dt as ClipboardChangedFilesPayloadSchema,At as ClipboardChangedPayloadSchema,wt as FILES_CHANGED,Zt as IMAGE_CHANGED,Tt as IS_MONITOR_RUNNING_COMMAND,Pt as MONITOR_UPDATE_EVENT,Ct as READ_FILES_COMMAND,Ot as READ_IMAGE_BINARY_COMMAND,It as READ_IMAGE_COMMAND,St as READ_TEXT_COMMAND,bt as START_MONITOR_COMMAND,xt as STOP_MONITOR_COMMAND,kt as TEXT_CHANGED,jt as WRITE_IMAGE_COMMAND,Et as WRITE_TEXT_COMMAND,$t as clear,Xt as isMonitorRunning,Ht as listenToClipboard,nn as listenToMonitorStatusUpdate,Gt as onClipboardUpdate,Yt as onFilesUpdate,Qt as onImageUpdate,Jt as onTextUpdate,Ut as readFiles,Wt as readImage,Vt as readImageBase64,zt as readImageBinary,Kt as readImageObjectURL,Mt as readText,qt as startBruteForceImageMonitor,Ft as startBruteForceTextMonitor,an as startListening,en as startMonitor,tn as stopMonitor,Bt as writeImage,Lt as writeText};
+function e(e,t,n,a){return new(n||(n=Promise))((function(r,s){function i(e){try{u(a.next(e))}catch(e){s(e)}}function o(e){try{u(a.throw(e))}catch(e){s(e)}}function u(e){var t;e.done?r(e.value):(t=e.value,t instanceof n?t:new n((function(e){e(t)}))).then(i,o)}u((a=a.apply(e,t||[])).next())}))}function t(e,t){var n,a,r,s,i={label:0,sent:function(){if(1&r[0])throw r[1];return r[1]},trys:[],ops:[]};return s={next:o(0),throw:o(1),return:o(2)},"function"==typeof Symbol&&(s[Symbol.iterator]=function(){return this}),s;function o(o){return function(u){return function(o){if(n)throw new TypeError("Generator is already executing.");for(;s&&(s=0,o[0]&&(i=0)),i;)try{if(n=1,a&&(r=2&o[0]?a.return:o[0]?a.throw||((r=a.return)&&r.call(a),0):a.next)&&!(r=r.call(a,o[1])).done)return r;switch(a=0,r&&(o=[2&o[0],r.value]),o[0]){case 0:case 1:r=o;break;case 4:return i.label++,{value:o[1],done:!1};case 5:i.label++,a=o[1],o=[0];continue;case 7:o=i.ops.pop(),i.trys.pop();continue;default:if(!(r=i.trys,(r=r.length>0&&r[r.length-1])||6!==o[0]&&2!==o[0])){i=0;continue}if(3===o[0]&&(!r||o[1]>r[0]&&o[1]<r[3])){i.label=o[1];break}if(6===o[0]&&i.label<r[1]){i.label=r[1],r=o;break}if(r&&i.label<r[2]){i.label=r[2],i.ops.push(o);break}r[2]&&i.ops.pop(),i.trys.pop();continue}o=t.call(e,i)}catch(e){o=[6,e],a=0}finally{n=r=0}if(5&o[0])throw o[1];return{value:o[0]?o[1]:void 0,done:!0}}([o,u])}}}var n,a;!function(e){e.assertEqual=e=>e,e.assertIs=function(e){},e.assertNever=function(e){throw new Error},e.arrayToEnum=e=>{const t={};for(const n of e)t[n]=n;return t},e.getValidEnumValues=t=>{const n=e.objectKeys(t).filter((e=>"number"!=typeof t[t[e]])),a={};for(const e of n)a[e]=t[e];return e.objectValues(a)},e.objectValues=t=>e.objectKeys(t).map((function(e){return t[e]})),e.objectKeys="function"==typeof Object.keys?e=>Object.keys(e):e=>{const t=[];for(const n in e)Object.prototype.hasOwnProperty.call(e,n)&&t.push(n);return t},e.find=(e,t)=>{for(const n of e)if(t(n))return n},e.isInteger="function"==typeof Number.isInteger?e=>Number.isInteger(e):e=>"number"==typeof e&&isFinite(e)&&Math.floor(e)===e,e.joinValues=function(e,t=" | "){return e.map((e=>"string"==typeof e?`'${e}'`:e)).join(t)},e.jsonStringifyReplacer=(e,t)=>"bigint"==typeof t?t.toString():t}(n||(n={})),function(e){e.mergeShapes=(e,t)=>({...e,...t})}(a||(a={}));const r=n.arrayToEnum(["string","nan","number","integer","float","boolean","date","bigint","symbol","function","undefined","null","array","object","unknown","promise","void","never","map","set"]),s=e=>{switch(typeof e){case"undefined":return r.undefined;case"string":return r.string;case"number":return isNaN(e)?r.nan:r.number;case"boolean":return r.boolean;case"function":return r.function;case"bigint":return r.bigint;case"symbol":return r.symbol;case"object":return Array.isArray(e)?r.array:null===e?r.null:e.then&&"function"==typeof e.then&&e.catch&&"function"==typeof e.catch?r.promise:"undefined"!=typeof Map&&e instanceof Map?r.map:"undefined"!=typeof Set&&e instanceof Set?r.set:"undefined"!=typeof Date&&e instanceof Date?r.date:r.object;default:return r.unknown}},i=n.arrayToEnum(["invalid_type","invalid_literal","custom","invalid_union","invalid_union_discriminator","invalid_enum_value","unrecognized_keys","invalid_arguments","invalid_return_type","invalid_date","invalid_string","too_small","too_big","invalid_intersection_types","not_multiple_of","not_finite"]);class o extends Error{constructor(e){super(),this.issues=[],this.addIssue=e=>{this.issues=[...this.issues,e]},this.addIssues=(e=[])=>{this.issues=[...this.issues,...e]};const t=new.target.prototype;Object.setPrototypeOf?Object.setPrototypeOf(this,t):this.__proto__=t,this.name="ZodError",this.issues=e}get errors(){return this.issues}format(e){const t=e||function(e){return e.message},n={_errors:[]},a=e=>{for(const r of e.issues)if("invalid_union"===r.code)r.unionErrors.map(a);else if("invalid_return_type"===r.code)a(r.returnTypeError);else if("invalid_arguments"===r.code)a(r.argumentsError);else if(0===r.path.length)n._errors.push(t(r));else{let e=n,a=0;for(;a<r.path.length;){const n=r.path[a];a===r.path.length-1?(e[n]=e[n]||{_errors:[]},e[n]._errors.push(t(r))):e[n]=e[n]||{_errors:[]},e=e[n],a++}}};return a(this),n}toString(){return this.message}get message(){return JSON.stringify(this.issues,n.jsonStringifyReplacer,2)}get isEmpty(){return 0===this.issues.length}flatten(e=(e=>e.message)){const t={},n=[];for(const a of this.issues)a.path.length>0?(t[a.path[0]]=t[a.path[0]]||[],t[a.path[0]].push(e(a))):n.push(e(a));return{formErrors:n,fieldErrors:t}}get formErrors(){return this.flatten()}}o.create=e=>new o(e);const u=(e,t)=>{let a;switch(e.code){case i.invalid_type:a=e.received===r.undefined?"Required":`Expected ${e.expected}, received ${e.received}`;break;case i.invalid_literal:a=`Invalid literal value, expected ${JSON.stringify(e.expected,n.jsonStringifyReplacer)}`;break;case i.unrecognized_keys:a=`Unrecognized key(s) in object: ${n.joinValues(e.keys,", ")}`;break;case i.invalid_union:a="Invalid input";break;case i.invalid_union_discriminator:a=`Invalid discriminator value. Expected ${n.joinValues(e.options)}`;break;case i.invalid_enum_value:a=`Invalid enum value. Expected ${n.joinValues(e.options)}, received '${e.received}'`;break;case i.invalid_arguments:a="Invalid function arguments";break;case i.invalid_return_type:a="Invalid function return type";break;case i.invalid_date:a="Invalid date";break;case i.invalid_string:"object"==typeof e.validation?"includes"in e.validation?(a=`Invalid input: must include "${e.validation.includes}"`,"number"==typeof e.validation.position&&(a=`${a} at one or more positions greater than or equal to ${e.validation.position}`)):"startsWith"in e.validation?a=`Invalid input: must start with "${e.validation.startsWith}"`:"endsWith"in e.validation?a=`Invalid input: must end with "${e.validation.endsWith}"`:n.assertNever(e.validation):a="regex"!==e.validation?`Invalid ${e.validation}`:"Invalid";break;case i.too_small:a="array"===e.type?`Array must contain ${e.exact?"exactly":e.inclusive?"at least":"more than"} ${e.minimum} element(s)`:"string"===e.type?`String must contain ${e.exact?"exactly":e.inclusive?"at least":"over"} ${e.minimum} character(s)`:"number"===e.type?`Number must be ${e.exact?"exactly equal to ":e.inclusive?"greater than or equal to ":"greater than "}${e.minimum}`:"date"===e.type?`Date must be ${e.exact?"exactly equal to ":e.inclusive?"greater than or equal to ":"greater than "}${new Date(Number(e.minimum))}`:"Invalid input";break;case i.too_big:a="array"===e.type?`Array must contain ${e.exact?"exactly":e.inclusive?"at most":"less than"} ${e.maximum} element(s)`:"string"===e.type?`String must contain ${e.exact?"exactly":e.inclusive?"at most":"under"} ${e.maximum} character(s)`:"number"===e.type?`Number must be ${e.exact?"exactly":e.inclusive?"less than or equal to":"less than"} ${e.maximum}`:"bigint"===e.type?`BigInt must be ${e.exact?"exactly":e.inclusive?"less than or equal to":"less than"} ${e.maximum}`:"date"===e.type?`Date must be ${e.exact?"exactly":e.inclusive?"smaller than or equal to":"smaller than"} ${new Date(Number(e.maximum))}`:"Invalid input";break;case i.custom:a="Invalid input";break;case i.invalid_intersection_types:a="Intersection results could not be merged";break;case i.not_multiple_of:a=`Number must be a multiple of ${e.multipleOf}`;break;case i.not_finite:a="Number must be finite";break;default:a=t.defaultError,n.assertNever(e)}return{message:a}};let d=u;function c(){return d}const l=e=>{const{data:t,path:n,errorMaps:a,issueData:r}=e,s=[...n,...r.path||[]],i={...r,path:s};let o="";const u=a.filter((e=>!!e)).slice().reverse();for(const e of u)o=e(i,{data:t,defaultError:o}).message;return{...r,path:s,message:r.message||o}};function p(e,t){const n=l({issueData:t,data:e.data,path:e.path,errorMaps:[e.common.contextualErrorMap,e.schemaErrorMap,c(),u].filter((e=>!!e))});e.common.issues.push(n)}class h{constructor(){this.value="valid"}dirty(){"valid"===this.value&&(this.value="dirty")}abort(){"aborted"!==this.value&&(this.value="aborted")}static mergeArray(e,t){const n=[];for(const a of t){if("aborted"===a.status)return f;"dirty"===a.status&&e.dirty(),n.push(a.value)}return{status:e.value,value:n}}static async mergeObjectAsync(e,t){const n=[];for(const e of t)n.push({key:await e.key,value:await e.value});return h.mergeObjectSync(e,n)}static mergeObjectSync(e,t){const n={};for(const a of t){const{key:t,value:r}=a;if("aborted"===t.status)return f;if("aborted"===r.status)return f;"dirty"===t.status&&e.dirty(),"dirty"===r.status&&e.dirty(),"__proto__"===t.value||void 0===r.value&&!a.alwaysSet||(n[t.value]=r.value)}return{status:e.value,value:n}}}const f=Object.freeze({status:"aborted"}),m=e=>({status:"dirty",value:e}),_=e=>({status:"valid",value:e}),y=e=>"aborted"===e.status,v=e=>"dirty"===e.status,g=e=>"valid"===e.status,b=e=>"undefined"!=typeof Promise&&e instanceof Promise;var x;!function(e){e.errToObj=e=>"string"==typeof e?{message:e}:e||{},e.toString=e=>"string"==typeof e?e:null==e?void 0:e.message}(x||(x={}));class k{constructor(e,t,n,a){this._cachedPath=[],this.parent=e,this.data=t,this._path=n,this._key=a}get path(){return this._cachedPath.length||(this._key instanceof Array?this._cachedPath.push(...this._path,...this._key):this._cachedPath.push(...this._path,this._key)),this._cachedPath}}const w=(e,t)=>{if(g(t))return{success:!0,data:t.value};if(!e.common.issues.length)throw new Error("Validation failed but no issues detected.");return{success:!1,get error(){if(this._error)return this._error;const t=new o(e.common.issues);return this._error=t,this._error}}};function T(e){if(!e)return{};const{errorMap:t,invalid_type_error:n,required_error:a,description:r}=e;if(t&&(n||a))throw new Error('Can\'t use "invalid_type_error" or "required_error" in conjunction with custom error map.');if(t)return{errorMap:t,description:r};return{errorMap:(e,t)=>"invalid_type"!==e.code?{message:t.defaultError}:void 0===t.data?{message:null!=a?a:t.defaultError}:{message:null!=n?n:t.defaultError},description:r}}class O{constructor(e){this.spa=this.safeParseAsync,this._def=e,this.parse=this.parse.bind(this),this.safeParse=this.safeParse.bind(this),this.parseAsync=this.parseAsync.bind(this),this.safeParseAsync=this.safeParseAsync.bind(this),this.spa=this.spa.bind(this),this.refine=this.refine.bind(this),this.refinement=this.refinement.bind(this),this.superRefine=this.superRefine.bind(this),this.optional=this.optional.bind(this),this.nullable=this.nullable.bind(this),this.nullish=this.nullish.bind(this),this.array=this.array.bind(this),this.promise=this.promise.bind(this),this.or=this.or.bind(this),this.and=this.and.bind(this),this.transform=this.transform.bind(this),this.brand=this.brand.bind(this),this.default=this.default.bind(this),this.catch=this.catch.bind(this),this.describe=this.describe.bind(this),this.pipe=this.pipe.bind(this),this.readonly=this.readonly.bind(this),this.isNullable=this.isNullable.bind(this),this.isOptional=this.isOptional.bind(this)}get description(){return this._def.description}_getType(e){return s(e.data)}_getOrReturnCtx(e,t){return t||{common:e.parent.common,data:e.data,parsedType:s(e.data),schemaErrorMap:this._def.errorMap,path:e.path,parent:e.parent}}_processInputParams(e){return{status:new h,ctx:{common:e.parent.common,data:e.data,parsedType:s(e.data),schemaErrorMap:this._def.errorMap,path:e.path,parent:e.parent}}}_parseSync(e){const t=this._parse(e);if(b(t))throw new Error("Synchronous parse encountered promise.");return t}_parseAsync(e){const t=this._parse(e);return Promise.resolve(t)}parse(e,t){const n=this.safeParse(e,t);if(n.success)return n.data;throw n.error}safeParse(e,t){var n;const a={common:{issues:[],async:null!==(n=null==t?void 0:t.async)&&void 0!==n&&n,contextualErrorMap:null==t?void 0:t.errorMap},path:(null==t?void 0:t.path)||[],schemaErrorMap:this._def.errorMap,parent:null,data:e,parsedType:s(e)},r=this._parseSync({data:e,path:a.path,parent:a});return w(a,r)}async parseAsync(e,t){const n=await this.safeParseAsync(e,t);if(n.success)return n.data;throw n.error}async safeParseAsync(e,t){const n={common:{issues:[],contextualErrorMap:null==t?void 0:t.errorMap,async:!0},path:(null==t?void 0:t.path)||[],schemaErrorMap:this._def.errorMap,parent:null,data:e,parsedType:s(e)},a=this._parse({data:e,path:n.path,parent:n}),r=await(b(a)?a:Promise.resolve(a));return w(n,r)}refine(e,t){const n=e=>"string"==typeof t||void 0===t?{message:t}:"function"==typeof t?t(e):t;return this._refinement(((t,a)=>{const r=e(t),s=()=>a.addIssue({code:i.custom,...n(t)});return"undefined"!=typeof Promise&&r instanceof Promise?r.then((e=>!!e||(s(),!1))):!!r||(s(),!1)}))}refinement(e,t){return this._refinement(((n,a)=>!!e(n)||(a.addIssue("function"==typeof t?t(n,a):t),!1)))}_refinement(e){return new pe({schema:this,typeName:Te.ZodEffects,effect:{type:"refinement",refinement:e}})}superRefine(e){return this._refinement(e)}optional(){return he.create(this,this._def)}nullable(){return fe.create(this,this._def)}nullish(){return this.nullable().optional()}array(){return H.create(this,this._def)}promise(){return le.create(this,this._def)}or(e){return Y.create([this,e],this._def)}and(e){return ee.create(this,e,this._def)}transform(e){return new pe({...T(this._def),schema:this,typeName:Te.ZodEffects,effect:{type:"transform",transform:e}})}default(e){const t="function"==typeof e?e:()=>e;return new me({...T(this._def),innerType:this,defaultValue:t,typeName:Te.ZodDefault})}brand(){return new ge({typeName:Te.ZodBranded,type:this,...T(this._def)})}catch(e){const t="function"==typeof e?e:()=>e;return new _e({...T(this._def),innerType:this,catchValue:t,typeName:Te.ZodCatch})}describe(e){return new(0,this.constructor)({...this._def,description:e})}pipe(e){return be.create(this,e)}readonly(){return xe.create(this)}isOptional(){return this.safeParse(void 0).success}isNullable(){return this.safeParse(null).success}}const Z=/^c[^\s-]{8,}$/i,E=/^[a-z][a-z0-9]*$/,N=/^[0-9A-HJKMNP-TV-Z]{26}$/,C=/^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$/i,A=/^(?!\.)(?!.*\.\.)([A-Z0-9_+-\.]*)[A-Z0-9_+-]@([A-Z0-9][A-Z0-9\-]*\.)+[A-Z]{2,}$/i;let I;const S=/^(((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2}))\.){3}((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2}))$/,R=/^(([a-f0-9]{1,4}:){7}|::([a-f0-9]{1,4}:){0,6}|([a-f0-9]{1,4}:){1}:([a-f0-9]{1,4}:){0,5}|([a-f0-9]{1,4}:){2}:([a-f0-9]{1,4}:){0,4}|([a-f0-9]{1,4}:){3}:([a-f0-9]{1,4}:){0,3}|([a-f0-9]{1,4}:){4}:([a-f0-9]{1,4}:){0,2}|([a-f0-9]{1,4}:){5}:([a-f0-9]{1,4}:){0,1})([a-f0-9]{1,4}|(((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2}))\.){3}((25[0-5])|(2[0-4][0-9])|(1[0-9]{2})|([0-9]{1,2})))$/;class M extends O{_parse(e){this._def.coerce&&(e.data=String(e.data));if(this._getType(e)!==r.string){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.string,received:t.parsedType}),f}const t=new h;let a;for(const r of this._def.checks)if("min"===r.kind)e.data.length<r.value&&(a=this._getOrReturnCtx(e,a),p(a,{code:i.too_small,minimum:r.value,type:"string",inclusive:!0,exact:!1,message:r.message}),t.dirty());else if("max"===r.kind)e.data.length>r.value&&(a=this._getOrReturnCtx(e,a),p(a,{code:i.too_big,maximum:r.value,type:"string",inclusive:!0,exact:!1,message:r.message}),t.dirty());else if("length"===r.kind){const n=e.data.length>r.value,s=e.data.length<r.value;(n||s)&&(a=this._getOrReturnCtx(e,a),n?p(a,{code:i.too_big,maximum:r.value,type:"string",inclusive:!0,exact:!0,message:r.message}):s&&p(a,{code:i.too_small,minimum:r.value,type:"string",inclusive:!0,exact:!0,message:r.message}),t.dirty())}else if("email"===r.kind)A.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"email",code:i.invalid_string,message:r.message}),t.dirty());else if("emoji"===r.kind)I||(I=new RegExp("^(\\p{Extended_Pictographic}|\\p{Emoji_Component})+$","u")),I.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"emoji",code:i.invalid_string,message:r.message}),t.dirty());else if("uuid"===r.kind)C.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"uuid",code:i.invalid_string,message:r.message}),t.dirty());else if("cuid"===r.kind)Z.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"cuid",code:i.invalid_string,message:r.message}),t.dirty());else if("cuid2"===r.kind)E.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"cuid2",code:i.invalid_string,message:r.message}),t.dirty());else if("ulid"===r.kind)N.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"ulid",code:i.invalid_string,message:r.message}),t.dirty());else if("url"===r.kind)try{new URL(e.data)}catch(n){a=this._getOrReturnCtx(e,a),p(a,{validation:"url",code:i.invalid_string,message:r.message}),t.dirty()}else if("regex"===r.kind){r.regex.lastIndex=0;r.regex.test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{validation:"regex",code:i.invalid_string,message:r.message}),t.dirty())}else if("trim"===r.kind)e.data=e.data.trim();else if("includes"===r.kind)e.data.includes(r.value,r.position)||(a=this._getOrReturnCtx(e,a),p(a,{code:i.invalid_string,validation:{includes:r.value,position:r.position},message:r.message}),t.dirty());else if("toLowerCase"===r.kind)e.data=e.data.toLowerCase();else if("toUpperCase"===r.kind)e.data=e.data.toUpperCase();else if("startsWith"===r.kind)e.data.startsWith(r.value)||(a=this._getOrReturnCtx(e,a),p(a,{code:i.invalid_string,validation:{startsWith:r.value},message:r.message}),t.dirty());else if("endsWith"===r.kind)e.data.endsWith(r.value)||(a=this._getOrReturnCtx(e,a),p(a,{code:i.invalid_string,validation:{endsWith:r.value},message:r.message}),t.dirty());else if("datetime"===r.kind){((u=r).precision?u.offset?new RegExp(`^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{${u.precision}}(([+-]\\d{2}(:?\\d{2})?)|Z)$`):new RegExp(`^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{${u.precision}}Z$`):0===u.precision?u.offset?new RegExp("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(([+-]\\d{2}(:?\\d{2})?)|Z)$"):new RegExp("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"):u.offset?new RegExp("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d+)?(([+-]\\d{2}(:?\\d{2})?)|Z)$"):new RegExp("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d+)?Z$")).test(e.data)||(a=this._getOrReturnCtx(e,a),p(a,{code:i.invalid_string,validation:"datetime",message:r.message}),t.dirty())}else"ip"===r.kind?(s=e.data,("v4"!==(o=r.version)&&o||!S.test(s))&&("v6"!==o&&o||!R.test(s))&&(a=this._getOrReturnCtx(e,a),p(a,{validation:"ip",code:i.invalid_string,message:r.message}),t.dirty())):n.assertNever(r);var s,o,u;return{status:t.value,value:e.data}}_regex(e,t,n){return this.refinement((t=>e.test(t)),{validation:t,code:i.invalid_string,...x.errToObj(n)})}_addCheck(e){return new M({...this._def,checks:[...this._def.checks,e]})}email(e){return this._addCheck({kind:"email",...x.errToObj(e)})}url(e){return this._addCheck({kind:"url",...x.errToObj(e)})}emoji(e){return this._addCheck({kind:"emoji",...x.errToObj(e)})}uuid(e){return this._addCheck({kind:"uuid",...x.errToObj(e)})}cuid(e){return this._addCheck({kind:"cuid",...x.errToObj(e)})}cuid2(e){return this._addCheck({kind:"cuid2",...x.errToObj(e)})}ulid(e){return this._addCheck({kind:"ulid",...x.errToObj(e)})}ip(e){return this._addCheck({kind:"ip",...x.errToObj(e)})}datetime(e){var t;return"string"==typeof e?this._addCheck({kind:"datetime",precision:null,offset:!1,message:e}):this._addCheck({kind:"datetime",precision:void 0===(null==e?void 0:e.precision)?null:null==e?void 0:e.precision,offset:null!==(t=null==e?void 0:e.offset)&&void 0!==t&&t,...x.errToObj(null==e?void 0:e.message)})}regex(e,t){return this._addCheck({kind:"regex",regex:e,...x.errToObj(t)})}includes(e,t){return this._addCheck({kind:"includes",value:e,position:null==t?void 0:t.position,...x.errToObj(null==t?void 0:t.message)})}startsWith(e,t){return this._addCheck({kind:"startsWith",value:e,...x.errToObj(t)})}endsWith(e,t){return this._addCheck({kind:"endsWith",value:e,...x.errToObj(t)})}min(e,t){return this._addCheck({kind:"min",value:e,...x.errToObj(t)})}max(e,t){return this._addCheck({kind:"max",value:e,...x.errToObj(t)})}length(e,t){return this._addCheck({kind:"length",value:e,...x.errToObj(t)})}nonempty(e){return this.min(1,x.errToObj(e))}trim(){return new M({...this._def,checks:[...this._def.checks,{kind:"trim"}]})}toLowerCase(){return new M({...this._def,checks:[...this._def.checks,{kind:"toLowerCase"}]})}toUpperCase(){return new M({...this._def,checks:[...this._def.checks,{kind:"toUpperCase"}]})}get isDatetime(){return!!this._def.checks.find((e=>"datetime"===e.kind))}get isEmail(){return!!this._def.checks.find((e=>"email"===e.kind))}get isURL(){return!!this._def.checks.find((e=>"url"===e.kind))}get isEmoji(){return!!this._def.checks.find((e=>"emoji"===e.kind))}get isUUID(){return!!this._def.checks.find((e=>"uuid"===e.kind))}get isCUID(){return!!this._def.checks.find((e=>"cuid"===e.kind))}get isCUID2(){return!!this._def.checks.find((e=>"cuid2"===e.kind))}get isULID(){return!!this._def.checks.find((e=>"ulid"===e.kind))}get isIP(){return!!this._def.checks.find((e=>"ip"===e.kind))}get minLength(){let e=null;for(const t of this._def.checks)"min"===t.kind&&(null===e||t.value>e)&&(e=t.value);return e}get maxLength(){let e=null;for(const t of this._def.checks)"max"===t.kind&&(null===e||t.value<e)&&(e=t.value);return e}}function D(e,t){const n=(e.toString().split(".")[1]||"").length,a=(t.toString().split(".")[1]||"").length,r=n>a?n:a;return parseInt(e.toFixed(r).replace(".",""))%parseInt(t.toFixed(r).replace(".",""))/Math.pow(10,r)}M.create=e=>{var t;return new M({checks:[],typeName:Te.ZodString,coerce:null!==(t=null==e?void 0:e.coerce)&&void 0!==t&&t,...T(e)})};class j extends O{constructor(){super(...arguments),this.min=this.gte,this.max=this.lte,this.step=this.multipleOf}_parse(e){this._def.coerce&&(e.data=Number(e.data));if(this._getType(e)!==r.number){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.number,received:t.parsedType}),f}let t;const a=new h;for(const r of this._def.checks)if("int"===r.kind)n.isInteger(e.data)||(t=this._getOrReturnCtx(e,t),p(t,{code:i.invalid_type,expected:"integer",received:"float",message:r.message}),a.dirty());else if("min"===r.kind){(r.inclusive?e.data<r.value:e.data<=r.value)&&(t=this._getOrReturnCtx(e,t),p(t,{code:i.too_small,minimum:r.value,type:"number",inclusive:r.inclusive,exact:!1,message:r.message}),a.dirty())}else if("max"===r.kind){(r.inclusive?e.data>r.value:e.data>=r.value)&&(t=this._getOrReturnCtx(e,t),p(t,{code:i.too_big,maximum:r.value,type:"number",inclusive:r.inclusive,exact:!1,message:r.message}),a.dirty())}else"multipleOf"===r.kind?0!==D(e.data,r.value)&&(t=this._getOrReturnCtx(e,t),p(t,{code:i.not_multiple_of,multipleOf:r.value,message:r.message}),a.dirty()):"finite"===r.kind?Number.isFinite(e.data)||(t=this._getOrReturnCtx(e,t),p(t,{code:i.not_finite,message:r.message}),a.dirty()):n.assertNever(r);return{status:a.value,value:e.data}}gte(e,t){return this.setLimit("min",e,!0,x.toString(t))}gt(e,t){return this.setLimit("min",e,!1,x.toString(t))}lte(e,t){return this.setLimit("max",e,!0,x.toString(t))}lt(e,t){return this.setLimit("max",e,!1,x.toString(t))}setLimit(e,t,n,a){return new j({...this._def,checks:[...this._def.checks,{kind:e,value:t,inclusive:n,message:x.toString(a)}]})}_addCheck(e){return new j({...this._def,checks:[...this._def.checks,e]})}int(e){return this._addCheck({kind:"int",message:x.toString(e)})}positive(e){return this._addCheck({kind:"min",value:0,inclusive:!1,message:x.toString(e)})}negative(e){return this._addCheck({kind:"max",value:0,inclusive:!1,message:x.toString(e)})}nonpositive(e){return this._addCheck({kind:"max",value:0,inclusive:!0,message:x.toString(e)})}nonnegative(e){return this._addCheck({kind:"min",value:0,inclusive:!0,message:x.toString(e)})}multipleOf(e,t){return this._addCheck({kind:"multipleOf",value:e,message:x.toString(t)})}finite(e){return this._addCheck({kind:"finite",message:x.toString(e)})}safe(e){return this._addCheck({kind:"min",inclusive:!0,value:Number.MIN_SAFE_INTEGER,message:x.toString(e)})._addCheck({kind:"max",inclusive:!0,value:Number.MAX_SAFE_INTEGER,message:x.toString(e)})}get minValue(){let e=null;for(const t of this._def.checks)"min"===t.kind&&(null===e||t.value>e)&&(e=t.value);return e}get maxValue(){let e=null;for(const t of this._def.checks)"max"===t.kind&&(null===e||t.value<e)&&(e=t.value);return e}get isInt(){return!!this._def.checks.find((e=>"int"===e.kind||"multipleOf"===e.kind&&n.isInteger(e.value)))}get isFinite(){let e=null,t=null;for(const n of this._def.checks){if("finite"===n.kind||"int"===n.kind||"multipleOf"===n.kind)return!0;"min"===n.kind?(null===t||n.value>t)&&(t=n.value):"max"===n.kind&&(null===e||n.value<e)&&(e=n.value)}return Number.isFinite(t)&&Number.isFinite(e)}}j.create=e=>new j({checks:[],typeName:Te.ZodNumber,coerce:(null==e?void 0:e.coerce)||!1,...T(e)});class P extends O{constructor(){super(...arguments),this.min=this.gte,this.max=this.lte}_parse(e){this._def.coerce&&(e.data=BigInt(e.data));if(this._getType(e)!==r.bigint){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.bigint,received:t.parsedType}),f}let t;const a=new h;for(const r of this._def.checks)if("min"===r.kind){(r.inclusive?e.data<r.value:e.data<=r.value)&&(t=this._getOrReturnCtx(e,t),p(t,{code:i.too_small,type:"bigint",minimum:r.value,inclusive:r.inclusive,message:r.message}),a.dirty())}else if("max"===r.kind){(r.inclusive?e.data>r.value:e.data>=r.value)&&(t=this._getOrReturnCtx(e,t),p(t,{code:i.too_big,type:"bigint",maximum:r.value,inclusive:r.inclusive,message:r.message}),a.dirty())}else"multipleOf"===r.kind?e.data%r.value!==BigInt(0)&&(t=this._getOrReturnCtx(e,t),p(t,{code:i.not_multiple_of,multipleOf:r.value,message:r.message}),a.dirty()):n.assertNever(r);return{status:a.value,value:e.data}}gte(e,t){return this.setLimit("min",e,!0,x.toString(t))}gt(e,t){return this.setLimit("min",e,!1,x.toString(t))}lte(e,t){return this.setLimit("max",e,!0,x.toString(t))}lt(e,t){return this.setLimit("max",e,!1,x.toString(t))}setLimit(e,t,n,a){return new P({...this._def,checks:[...this._def.checks,{kind:e,value:t,inclusive:n,message:x.toString(a)}]})}_addCheck(e){return new P({...this._def,checks:[...this._def.checks,e]})}positive(e){return this._addCheck({kind:"min",value:BigInt(0),inclusive:!1,message:x.toString(e)})}negative(e){return this._addCheck({kind:"max",value:BigInt(0),inclusive:!1,message:x.toString(e)})}nonpositive(e){return this._addCheck({kind:"max",value:BigInt(0),inclusive:!0,message:x.toString(e)})}nonnegative(e){return this._addCheck({kind:"min",value:BigInt(0),inclusive:!0,message:x.toString(e)})}multipleOf(e,t){return this._addCheck({kind:"multipleOf",value:e,message:x.toString(t)})}get minValue(){let e=null;for(const t of this._def.checks)"min"===t.kind&&(null===e||t.value>e)&&(e=t.value);return e}get maxValue(){let e=null;for(const t of this._def.checks)"max"===t.kind&&(null===e||t.value<e)&&(e=t.value);return e}}P.create=e=>{var t;return new P({checks:[],typeName:Te.ZodBigInt,coerce:null!==(t=null==e?void 0:e.coerce)&&void 0!==t&&t,...T(e)})};class L extends O{_parse(e){this._def.coerce&&(e.data=Boolean(e.data));if(this._getType(e)!==r.boolean){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.boolean,received:t.parsedType}),f}return _(e.data)}}L.create=e=>new L({typeName:Te.ZodBoolean,coerce:(null==e?void 0:e.coerce)||!1,...T(e)});class U extends O{_parse(e){this._def.coerce&&(e.data=new Date(e.data));if(this._getType(e)!==r.date){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.date,received:t.parsedType}),f}if(isNaN(e.data.getTime())){return p(this._getOrReturnCtx(e),{code:i.invalid_date}),f}const t=new h;let a;for(const r of this._def.checks)"min"===r.kind?e.data.getTime()<r.value&&(a=this._getOrReturnCtx(e,a),p(a,{code:i.too_small,message:r.message,inclusive:!0,exact:!1,minimum:r.value,type:"date"}),t.dirty()):"max"===r.kind?e.data.getTime()>r.value&&(a=this._getOrReturnCtx(e,a),p(a,{code:i.too_big,message:r.message,inclusive:!0,exact:!1,maximum:r.value,type:"date"}),t.dirty()):n.assertNever(r);return{status:t.value,value:new Date(e.data.getTime())}}_addCheck(e){return new U({...this._def,checks:[...this._def.checks,e]})}min(e,t){return this._addCheck({kind:"min",value:e.getTime(),message:x.toString(t)})}max(e,t){return this._addCheck({kind:"max",value:e.getTime(),message:x.toString(t)})}get minDate(){let e=null;for(const t of this._def.checks)"min"===t.kind&&(null===e||t.value>e)&&(e=t.value);return null!=e?new Date(e):null}get maxDate(){let e=null;for(const t of this._def.checks)"max"===t.kind&&(null===e||t.value<e)&&(e=t.value);return null!=e?new Date(e):null}}U.create=e=>new U({checks:[],coerce:(null==e?void 0:e.coerce)||!1,typeName:Te.ZodDate,...T(e)});class $ extends O{_parse(e){if(this._getType(e)!==r.symbol){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.symbol,received:t.parsedType}),f}return _(e.data)}}$.create=e=>new $({typeName:Te.ZodSymbol,...T(e)});class W extends O{_parse(e){if(this._getType(e)!==r.undefined){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.undefined,received:t.parsedType}),f}return _(e.data)}}W.create=e=>new W({typeName:Te.ZodUndefined,...T(e)});class V extends O{_parse(e){if(this._getType(e)!==r.null){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.null,received:t.parsedType}),f}return _(e.data)}}V.create=e=>new V({typeName:Te.ZodNull,...T(e)});class F extends O{constructor(){super(...arguments),this._any=!0}_parse(e){return _(e.data)}}F.create=e=>new F({typeName:Te.ZodAny,...T(e)});class B extends O{constructor(){super(...arguments),this._unknown=!0}_parse(e){return _(e.data)}}B.create=e=>new B({typeName:Te.ZodUnknown,...T(e)});class z extends O{_parse(e){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.never,received:t.parsedType}),f}}z.create=e=>new z({typeName:Te.ZodNever,...T(e)});class K extends O{_parse(e){if(this._getType(e)!==r.undefined){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.void,received:t.parsedType}),f}return _(e.data)}}K.create=e=>new K({typeName:Te.ZodVoid,...T(e)});class H extends O{_parse(e){const{ctx:t,status:n}=this._processInputParams(e),a=this._def;if(t.parsedType!==r.array)return p(t,{code:i.invalid_type,expected:r.array,received:t.parsedType}),f;if(null!==a.exactLength){const e=t.data.length>a.exactLength.value,r=t.data.length<a.exactLength.value;(e||r)&&(p(t,{code:e?i.too_big:i.too_small,minimum:r?a.exactLength.value:void 0,maximum:e?a.exactLength.value:void 0,type:"array",inclusive:!0,exact:!0,message:a.exactLength.message}),n.dirty())}if(null!==a.minLength&&t.data.length<a.minLength.value&&(p(t,{code:i.too_small,minimum:a.minLength.value,type:"array",inclusive:!0,exact:!1,message:a.minLength.message}),n.dirty()),null!==a.maxLength&&t.data.length>a.maxLength.value&&(p(t,{code:i.too_big,maximum:a.maxLength.value,type:"array",inclusive:!0,exact:!1,message:a.maxLength.message}),n.dirty()),t.common.async)return Promise.all([...t.data].map(((e,n)=>a.type._parseAsync(new k(t,e,t.path,n))))).then((e=>h.mergeArray(n,e)));const s=[...t.data].map(((e,n)=>a.type._parseSync(new k(t,e,t.path,n))));return h.mergeArray(n,s)}get element(){return this._def.type}min(e,t){return new H({...this._def,minLength:{value:e,message:x.toString(t)}})}max(e,t){return new H({...this._def,maxLength:{value:e,message:x.toString(t)}})}length(e,t){return new H({...this._def,exactLength:{value:e,message:x.toString(t)}})}nonempty(e){return this.min(1,e)}}function G(e){if(e instanceof q){const t={};for(const n in e.shape){const a=e.shape[n];t[n]=he.create(G(a))}return new q({...e._def,shape:()=>t})}return e instanceof H?new H({...e._def,type:G(e.element)}):e instanceof he?he.create(G(e.unwrap())):e instanceof fe?fe.create(G(e.unwrap())):e instanceof te?te.create(e.items.map((e=>G(e)))):e}H.create=(e,t)=>new H({type:e,minLength:null,maxLength:null,exactLength:null,typeName:Te.ZodArray,...T(t)});class q extends O{constructor(){super(...arguments),this._cached=null,this.nonstrict=this.passthrough,this.augment=this.extend}_getCached(){if(null!==this._cached)return this._cached;const e=this._def.shape(),t=n.objectKeys(e);return this._cached={shape:e,keys:t}}_parse(e){if(this._getType(e)!==r.object){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.object,received:t.parsedType}),f}const{status:t,ctx:n}=this._processInputParams(e),{shape:a,keys:s}=this._getCached(),o=[];if(!(this._def.catchall instanceof z&&"strip"===this._def.unknownKeys))for(const e in n.data)s.includes(e)||o.push(e);const u=[];for(const e of s){const t=a[e],r=n.data[e];u.push({key:{status:"valid",value:e},value:t._parse(new k(n,r,n.path,e)),alwaysSet:e in n.data})}if(this._def.catchall instanceof z){const e=this._def.unknownKeys;if("passthrough"===e)for(const e of o)u.push({key:{status:"valid",value:e},value:{status:"valid",value:n.data[e]}});else if("strict"===e)o.length>0&&(p(n,{code:i.unrecognized_keys,keys:o}),t.dirty());else if("strip"!==e)throw new Error("Internal ZodObject error: invalid unknownKeys value.")}else{const e=this._def.catchall;for(const t of o){const a=n.data[t];u.push({key:{status:"valid",value:t},value:e._parse(new k(n,a,n.path,t)),alwaysSet:t in n.data})}}return n.common.async?Promise.resolve().then((async()=>{const e=[];for(const t of u){const n=await t.key;e.push({key:n,value:await t.value,alwaysSet:t.alwaysSet})}return e})).then((e=>h.mergeObjectSync(t,e))):h.mergeObjectSync(t,u)}get shape(){return this._def.shape()}strict(e){return x.errToObj,new q({...this._def,unknownKeys:"strict",...void 0!==e?{errorMap:(t,n)=>{var a,r,s,i;const o=null!==(s=null===(r=(a=this._def).errorMap)||void 0===r?void 0:r.call(a,t,n).message)&&void 0!==s?s:n.defaultError;return"unrecognized_keys"===t.code?{message:null!==(i=x.errToObj(e).message)&&void 0!==i?i:o}:{message:o}}}:{}})}strip(){return new q({...this._def,unknownKeys:"strip"})}passthrough(){return new q({...this._def,unknownKeys:"passthrough"})}extend(e){return new q({...this._def,shape:()=>({...this._def.shape(),...e})})}merge(e){return new q({unknownKeys:e._def.unknownKeys,catchall:e._def.catchall,shape:()=>({...this._def.shape(),...e._def.shape()}),typeName:Te.ZodObject})}setKey(e,t){return this.augment({[e]:t})}catchall(e){return new q({...this._def,catchall:e})}pick(e){const t={};return n.objectKeys(e).forEach((n=>{e[n]&&this.shape[n]&&(t[n]=this.shape[n])})),new q({...this._def,shape:()=>t})}omit(e){const t={};return n.objectKeys(this.shape).forEach((n=>{e[n]||(t[n]=this.shape[n])})),new q({...this._def,shape:()=>t})}deepPartial(){return G(this)}partial(e){const t={};return n.objectKeys(this.shape).forEach((n=>{const a=this.shape[n];e&&!e[n]?t[n]=a:t[n]=a.optional()})),new q({...this._def,shape:()=>t})}required(e){const t={};return n.objectKeys(this.shape).forEach((n=>{if(e&&!e[n])t[n]=this.shape[n];else{let e=this.shape[n];for(;e instanceof he;)e=e._def.innerType;t[n]=e}})),new q({...this._def,shape:()=>t})}keyof(){return ue(n.objectKeys(this.shape))}}q.create=(e,t)=>new q({shape:()=>e,unknownKeys:"strip",catchall:z.create(),typeName:Te.ZodObject,...T(t)}),q.strictCreate=(e,t)=>new q({shape:()=>e,unknownKeys:"strict",catchall:z.create(),typeName:Te.ZodObject,...T(t)}),q.lazycreate=(e,t)=>new q({shape:e,unknownKeys:"strip",catchall:z.create(),typeName:Te.ZodObject,...T(t)});class Y extends O{_parse(e){const{ctx:t}=this._processInputParams(e),n=this._def.options;if(t.common.async)return Promise.all(n.map((async e=>{const n={...t,common:{...t.common,issues:[]},parent:null};return{result:await e._parseAsync({data:t.data,path:t.path,parent:n}),ctx:n}}))).then((function(e){for(const t of e)if("valid"===t.result.status)return t.result;for(const n of e)if("dirty"===n.result.status)return t.common.issues.push(...n.ctx.common.issues),n.result;const n=e.map((e=>new o(e.ctx.common.issues)));return p(t,{code:i.invalid_union,unionErrors:n}),f}));{let e;const a=[];for(const r of n){const n={...t,common:{...t.common,issues:[]},parent:null},s=r._parseSync({data:t.data,path:t.path,parent:n});if("valid"===s.status)return s;"dirty"!==s.status||e||(e={result:s,ctx:n}),n.common.issues.length&&a.push(n.common.issues)}if(e)return t.common.issues.push(...e.ctx.common.issues),e.result;const r=a.map((e=>new o(e)));return p(t,{code:i.invalid_union,unionErrors:r}),f}}get options(){return this._def.options}}Y.create=(e,t)=>new Y({options:e,typeName:Te.ZodUnion,...T(t)});const J=e=>e instanceof ie?J(e.schema):e instanceof pe?J(e.innerType()):e instanceof oe?[e.value]:e instanceof de?e.options:e instanceof ce?Object.keys(e.enum):e instanceof me?J(e._def.innerType):e instanceof W?[void 0]:e instanceof V?[null]:null;class X extends O{_parse(e){const{ctx:t}=this._processInputParams(e);if(t.parsedType!==r.object)return p(t,{code:i.invalid_type,expected:r.object,received:t.parsedType}),f;const n=this.discriminator,a=t.data[n],s=this.optionsMap.get(a);return s?t.common.async?s._parseAsync({data:t.data,path:t.path,parent:t}):s._parseSync({data:t.data,path:t.path,parent:t}):(p(t,{code:i.invalid_union_discriminator,options:Array.from(this.optionsMap.keys()),path:[n]}),f)}get discriminator(){return this._def.discriminator}get options(){return this._def.options}get optionsMap(){return this._def.optionsMap}static create(e,t,n){const a=new Map;for(const n of t){const t=J(n.shape[e]);if(!t)throw new Error(`A discriminator value for key \`${e}\` could not be extracted from all schema options`);for(const r of t){if(a.has(r))throw new Error(`Discriminator property ${String(e)} has duplicate value ${String(r)}`);a.set(r,n)}}return new X({typeName:Te.ZodDiscriminatedUnion,discriminator:e,options:t,optionsMap:a,...T(n)})}}function Q(e,t){const a=s(e),i=s(t);if(e===t)return{valid:!0,data:e};if(a===r.object&&i===r.object){const a=n.objectKeys(t),r=n.objectKeys(e).filter((e=>-1!==a.indexOf(e))),s={...e,...t};for(const n of r){const a=Q(e[n],t[n]);if(!a.valid)return{valid:!1};s[n]=a.data}return{valid:!0,data:s}}if(a===r.array&&i===r.array){if(e.length!==t.length)return{valid:!1};const n=[];for(let a=0;a<e.length;a++){const r=Q(e[a],t[a]);if(!r.valid)return{valid:!1};n.push(r.data)}return{valid:!0,data:n}}return a===r.date&&i===r.date&&+e==+t?{valid:!0,data:e}:{valid:!1}}class ee extends O{_parse(e){const{status:t,ctx:n}=this._processInputParams(e),a=(e,a)=>{if(y(e)||y(a))return f;const r=Q(e.value,a.value);return r.valid?((v(e)||v(a))&&t.dirty(),{status:t.value,value:r.data}):(p(n,{code:i.invalid_intersection_types}),f)};return n.common.async?Promise.all([this._def.left._parseAsync({data:n.data,path:n.path,parent:n}),this._def.right._parseAsync({data:n.data,path:n.path,parent:n})]).then((([e,t])=>a(e,t))):a(this._def.left._parseSync({data:n.data,path:n.path,parent:n}),this._def.right._parseSync({data:n.data,path:n.path,parent:n}))}}ee.create=(e,t,n)=>new ee({left:e,right:t,typeName:Te.ZodIntersection,...T(n)});class te extends O{_parse(e){const{status:t,ctx:n}=this._processInputParams(e);if(n.parsedType!==r.array)return p(n,{code:i.invalid_type,expected:r.array,received:n.parsedType}),f;if(n.data.length<this._def.items.length)return p(n,{code:i.too_small,minimum:this._def.items.length,inclusive:!0,exact:!1,type:"array"}),f;!this._def.rest&&n.data.length>this._def.items.length&&(p(n,{code:i.too_big,maximum:this._def.items.length,inclusive:!0,exact:!1,type:"array"}),t.dirty());const a=[...n.data].map(((e,t)=>{const a=this._def.items[t]||this._def.rest;return a?a._parse(new k(n,e,n.path,t)):null})).filter((e=>!!e));return n.common.async?Promise.all(a).then((e=>h.mergeArray(t,e))):h.mergeArray(t,a)}get items(){return this._def.items}rest(e){return new te({...this._def,rest:e})}}te.create=(e,t)=>{if(!Array.isArray(e))throw new Error("You must pass an array of schemas to z.tuple([ ... ])");return new te({items:e,typeName:Te.ZodTuple,rest:null,...T(t)})};class ne extends O{get keySchema(){return this._def.keyType}get valueSchema(){return this._def.valueType}_parse(e){const{status:t,ctx:n}=this._processInputParams(e);if(n.parsedType!==r.object)return p(n,{code:i.invalid_type,expected:r.object,received:n.parsedType}),f;const a=[],s=this._def.keyType,o=this._def.valueType;for(const e in n.data)a.push({key:s._parse(new k(n,e,n.path,e)),value:o._parse(new k(n,n.data[e],n.path,e))});return n.common.async?h.mergeObjectAsync(t,a):h.mergeObjectSync(t,a)}get element(){return this._def.valueType}static create(e,t,n){return new ne(t instanceof O?{keyType:e,valueType:t,typeName:Te.ZodRecord,...T(n)}:{keyType:M.create(),valueType:e,typeName:Te.ZodRecord,...T(t)})}}class ae extends O{get keySchema(){return this._def.keyType}get valueSchema(){return this._def.valueType}_parse(e){const{status:t,ctx:n}=this._processInputParams(e);if(n.parsedType!==r.map)return p(n,{code:i.invalid_type,expected:r.map,received:n.parsedType}),f;const a=this._def.keyType,s=this._def.valueType,o=[...n.data.entries()].map((([e,t],r)=>({key:a._parse(new k(n,e,n.path,[r,"key"])),value:s._parse(new k(n,t,n.path,[r,"value"]))})));if(n.common.async){const e=new Map;return Promise.resolve().then((async()=>{for(const n of o){const a=await n.key,r=await n.value;if("aborted"===a.status||"aborted"===r.status)return f;"dirty"!==a.status&&"dirty"!==r.status||t.dirty(),e.set(a.value,r.value)}return{status:t.value,value:e}}))}{const e=new Map;for(const n of o){const a=n.key,r=n.value;if("aborted"===a.status||"aborted"===r.status)return f;"dirty"!==a.status&&"dirty"!==r.status||t.dirty(),e.set(a.value,r.value)}return{status:t.value,value:e}}}}ae.create=(e,t,n)=>new ae({valueType:t,keyType:e,typeName:Te.ZodMap,...T(n)});class re extends O{_parse(e){const{status:t,ctx:n}=this._processInputParams(e);if(n.parsedType!==r.set)return p(n,{code:i.invalid_type,expected:r.set,received:n.parsedType}),f;const a=this._def;null!==a.minSize&&n.data.size<a.minSize.value&&(p(n,{code:i.too_small,minimum:a.minSize.value,type:"set",inclusive:!0,exact:!1,message:a.minSize.message}),t.dirty()),null!==a.maxSize&&n.data.size>a.maxSize.value&&(p(n,{code:i.too_big,maximum:a.maxSize.value,type:"set",inclusive:!0,exact:!1,message:a.maxSize.message}),t.dirty());const s=this._def.valueType;function o(e){const n=new Set;for(const a of e){if("aborted"===a.status)return f;"dirty"===a.status&&t.dirty(),n.add(a.value)}return{status:t.value,value:n}}const u=[...n.data.values()].map(((e,t)=>s._parse(new k(n,e,n.path,t))));return n.common.async?Promise.all(u).then((e=>o(e))):o(u)}min(e,t){return new re({...this._def,minSize:{value:e,message:x.toString(t)}})}max(e,t){return new re({...this._def,maxSize:{value:e,message:x.toString(t)}})}size(e,t){return this.min(e,t).max(e,t)}nonempty(e){return this.min(1,e)}}re.create=(e,t)=>new re({valueType:e,minSize:null,maxSize:null,typeName:Te.ZodSet,...T(t)});class se extends O{constructor(){super(...arguments),this.validate=this.implement}_parse(e){const{ctx:t}=this._processInputParams(e);if(t.parsedType!==r.function)return p(t,{code:i.invalid_type,expected:r.function,received:t.parsedType}),f;function n(e,n){return l({data:e,path:t.path,errorMaps:[t.common.contextualErrorMap,t.schemaErrorMap,c(),u].filter((e=>!!e)),issueData:{code:i.invalid_arguments,argumentsError:n}})}function a(e,n){return l({data:e,path:t.path,errorMaps:[t.common.contextualErrorMap,t.schemaErrorMap,c(),u].filter((e=>!!e)),issueData:{code:i.invalid_return_type,returnTypeError:n}})}const s={errorMap:t.common.contextualErrorMap},d=t.data;if(this._def.returns instanceof le){const e=this;return _((async function(...t){const r=new o([]),i=await e._def.args.parseAsync(t,s).catch((e=>{throw r.addIssue(n(t,e)),r})),u=await Reflect.apply(d,this,i),c=await e._def.returns._def.type.parseAsync(u,s).catch((e=>{throw r.addIssue(a(u,e)),r}));return c}))}{const e=this;return _((function(...t){const r=e._def.args.safeParse(t,s);if(!r.success)throw new o([n(t,r.error)]);const i=Reflect.apply(d,this,r.data),u=e._def.returns.safeParse(i,s);if(!u.success)throw new o([a(i,u.error)]);return u.data}))}}parameters(){return this._def.args}returnType(){return this._def.returns}args(...e){return new se({...this._def,args:te.create(e).rest(B.create())})}returns(e){return new se({...this._def,returns:e})}implement(e){return this.parse(e)}strictImplement(e){return this.parse(e)}static create(e,t,n){return new se({args:e||te.create([]).rest(B.create()),returns:t||B.create(),typeName:Te.ZodFunction,...T(n)})}}class ie extends O{get schema(){return this._def.getter()}_parse(e){const{ctx:t}=this._processInputParams(e);return this._def.getter()._parse({data:t.data,path:t.path,parent:t})}}ie.create=(e,t)=>new ie({getter:e,typeName:Te.ZodLazy,...T(t)});class oe extends O{_parse(e){if(e.data!==this._def.value){const t=this._getOrReturnCtx(e);return p(t,{received:t.data,code:i.invalid_literal,expected:this._def.value}),f}return{status:"valid",value:e.data}}get value(){return this._def.value}}function ue(e,t){return new de({values:e,typeName:Te.ZodEnum,...T(t)})}oe.create=(e,t)=>new oe({value:e,typeName:Te.ZodLiteral,...T(t)});class de extends O{_parse(e){if("string"!=typeof e.data){const t=this._getOrReturnCtx(e),a=this._def.values;return p(t,{expected:n.joinValues(a),received:t.parsedType,code:i.invalid_type}),f}if(-1===this._def.values.indexOf(e.data)){const t=this._getOrReturnCtx(e),n=this._def.values;return p(t,{received:t.data,code:i.invalid_enum_value,options:n}),f}return _(e.data)}get options(){return this._def.values}get enum(){const e={};for(const t of this._def.values)e[t]=t;return e}get Values(){const e={};for(const t of this._def.values)e[t]=t;return e}get Enum(){const e={};for(const t of this._def.values)e[t]=t;return e}extract(e){return de.create(e)}exclude(e){return de.create(this.options.filter((t=>!e.includes(t))))}}de.create=ue;class ce extends O{_parse(e){const t=n.getValidEnumValues(this._def.values),a=this._getOrReturnCtx(e);if(a.parsedType!==r.string&&a.parsedType!==r.number){const e=n.objectValues(t);return p(a,{expected:n.joinValues(e),received:a.parsedType,code:i.invalid_type}),f}if(-1===t.indexOf(e.data)){const e=n.objectValues(t);return p(a,{received:a.data,code:i.invalid_enum_value,options:e}),f}return _(e.data)}get enum(){return this._def.values}}ce.create=(e,t)=>new ce({values:e,typeName:Te.ZodNativeEnum,...T(t)});class le extends O{unwrap(){return this._def.type}_parse(e){const{ctx:t}=this._processInputParams(e);if(t.parsedType!==r.promise&&!1===t.common.async)return p(t,{code:i.invalid_type,expected:r.promise,received:t.parsedType}),f;const n=t.parsedType===r.promise?t.data:Promise.resolve(t.data);return _(n.then((e=>this._def.type.parseAsync(e,{path:t.path,errorMap:t.common.contextualErrorMap}))))}}le.create=(e,t)=>new le({type:e,typeName:Te.ZodPromise,...T(t)});class pe extends O{innerType(){return this._def.schema}sourceType(){return this._def.schema._def.typeName===Te.ZodEffects?this._def.schema.sourceType():this._def.schema}_parse(e){const{status:t,ctx:a}=this._processInputParams(e),r=this._def.effect||null,s={addIssue:e=>{p(a,e),e.fatal?t.abort():t.dirty()},get path(){return a.path}};if(s.addIssue=s.addIssue.bind(s),"preprocess"===r.type){const e=r.transform(a.data,s);return a.common.issues.length?{status:"dirty",value:a.data}:a.common.async?Promise.resolve(e).then((e=>this._def.schema._parseAsync({data:e,path:a.path,parent:a}))):this._def.schema._parseSync({data:e,path:a.path,parent:a})}if("refinement"===r.type){const e=e=>{const t=r.refinement(e,s);if(a.common.async)return Promise.resolve(t);if(t instanceof Promise)throw new Error("Async refinement encountered during synchronous parse operation. Use .parseAsync instead.");return e};if(!1===a.common.async){const n=this._def.schema._parseSync({data:a.data,path:a.path,parent:a});return"aborted"===n.status?f:("dirty"===n.status&&t.dirty(),e(n.value),{status:t.value,value:n.value})}return this._def.schema._parseAsync({data:a.data,path:a.path,parent:a}).then((n=>"aborted"===n.status?f:("dirty"===n.status&&t.dirty(),e(n.value).then((()=>({status:t.value,value:n.value}))))))}if("transform"===r.type){if(!1===a.common.async){const e=this._def.schema._parseSync({data:a.data,path:a.path,parent:a});if(!g(e))return e;const n=r.transform(e.value,s);if(n instanceof Promise)throw new Error("Asynchronous transform encountered during synchronous parse operation. Use .parseAsync instead.");return{status:t.value,value:n}}return this._def.schema._parseAsync({data:a.data,path:a.path,parent:a}).then((e=>g(e)?Promise.resolve(r.transform(e.value,s)).then((e=>({status:t.value,value:e}))):e))}n.assertNever(r)}}pe.create=(e,t,n)=>new pe({schema:e,typeName:Te.ZodEffects,effect:t,...T(n)}),pe.createWithPreprocess=(e,t,n)=>new pe({schema:t,effect:{type:"preprocess",transform:e},typeName:Te.ZodEffects,...T(n)});class he extends O{_parse(e){return this._getType(e)===r.undefined?_(void 0):this._def.innerType._parse(e)}unwrap(){return this._def.innerType}}he.create=(e,t)=>new he({innerType:e,typeName:Te.ZodOptional,...T(t)});class fe extends O{_parse(e){return this._getType(e)===r.null?_(null):this._def.innerType._parse(e)}unwrap(){return this._def.innerType}}fe.create=(e,t)=>new fe({innerType:e,typeName:Te.ZodNullable,...T(t)});class me extends O{_parse(e){const{ctx:t}=this._processInputParams(e);let n=t.data;return t.parsedType===r.undefined&&(n=this._def.defaultValue()),this._def.innerType._parse({data:n,path:t.path,parent:t})}removeDefault(){return this._def.innerType}}me.create=(e,t)=>new me({innerType:e,typeName:Te.ZodDefault,defaultValue:"function"==typeof t.default?t.default:()=>t.default,...T(t)});class _e extends O{_parse(e){const{ctx:t}=this._processInputParams(e),n={...t,common:{...t.common,issues:[]}},a=this._def.innerType._parse({data:n.data,path:n.path,parent:{...n}});return b(a)?a.then((e=>({status:"valid",value:"valid"===e.status?e.value:this._def.catchValue({get error(){return new o(n.common.issues)},input:n.data})}))):{status:"valid",value:"valid"===a.status?a.value:this._def.catchValue({get error(){return new o(n.common.issues)},input:n.data})}}removeCatch(){return this._def.innerType}}_e.create=(e,t)=>new _e({innerType:e,typeName:Te.ZodCatch,catchValue:"function"==typeof t.catch?t.catch:()=>t.catch,...T(t)});class ye extends O{_parse(e){if(this._getType(e)!==r.nan){const t=this._getOrReturnCtx(e);return p(t,{code:i.invalid_type,expected:r.nan,received:t.parsedType}),f}return{status:"valid",value:e.data}}}ye.create=e=>new ye({typeName:Te.ZodNaN,...T(e)});const ve=Symbol("zod_brand");class ge extends O{_parse(e){const{ctx:t}=this._processInputParams(e),n=t.data;return this._def.type._parse({data:n,path:t.path,parent:t})}unwrap(){return this._def.type}}class be extends O{_parse(e){const{status:t,ctx:n}=this._processInputParams(e);if(n.common.async){return(async()=>{const e=await this._def.in._parseAsync({data:n.data,path:n.path,parent:n});return"aborted"===e.status?f:"dirty"===e.status?(t.dirty(),m(e.value)):this._def.out._parseAsync({data:e.value,path:n.path,parent:n})})()}{const e=this._def.in._parseSync({data:n.data,path:n.path,parent:n});return"aborted"===e.status?f:"dirty"===e.status?(t.dirty(),{status:"dirty",value:e.value}):this._def.out._parseSync({data:e.value,path:n.path,parent:n})}}static create(e,t){return new be({in:e,out:t,typeName:Te.ZodPipeline})}}class xe extends O{_parse(e){const t=this._def.innerType._parse(e);return g(t)&&(t.value=Object.freeze(t.value)),t}}xe.create=(e,t)=>new xe({innerType:e,typeName:Te.ZodReadonly,...T(t)});const ke=(e,t={},n)=>e?F.create().superRefine(((a,r)=>{var s,i;if(!e(a)){const e="function"==typeof t?t(a):"string"==typeof t?{message:t}:t,o=null===(i=null!==(s=e.fatal)&&void 0!==s?s:n)||void 0===i||i,u="string"==typeof e?{message:e}:e;r.addIssue({code:"custom",...u,fatal:o})}})):F.create(),we={object:q.lazycreate};var Te;!function(e){e.ZodString="ZodString",e.ZodNumber="ZodNumber",e.ZodNaN="ZodNaN",e.ZodBigInt="ZodBigInt",e.ZodBoolean="ZodBoolean",e.ZodDate="ZodDate",e.ZodSymbol="ZodSymbol",e.ZodUndefined="ZodUndefined",e.ZodNull="ZodNull",e.ZodAny="ZodAny",e.ZodUnknown="ZodUnknown",e.ZodNever="ZodNever",e.ZodVoid="ZodVoid",e.ZodArray="ZodArray",e.ZodObject="ZodObject",e.ZodUnion="ZodUnion",e.ZodDiscriminatedUnion="ZodDiscriminatedUnion",e.ZodIntersection="ZodIntersection",e.ZodTuple="ZodTuple",e.ZodRecord="ZodRecord",e.ZodMap="ZodMap",e.ZodSet="ZodSet",e.ZodFunction="ZodFunction",e.ZodLazy="ZodLazy",e.ZodLiteral="ZodLiteral",e.ZodEnum="ZodEnum",e.ZodEffects="ZodEffects",e.ZodNativeEnum="ZodNativeEnum",e.ZodOptional="ZodOptional",e.ZodNullable="ZodNullable",e.ZodDefault="ZodDefault",e.ZodCatch="ZodCatch",e.ZodPromise="ZodPromise",e.ZodBranded="ZodBranded",e.ZodPipeline="ZodPipeline",e.ZodReadonly="ZodReadonly"}(Te||(Te={}));const Oe=M.create,Ze=j.create,Ee=ye.create,Ne=P.create,Ce=L.create,Ae=U.create,Ie=$.create,Se=W.create,Re=V.create,Me=F.create,De=B.create,je=z.create,Pe=K.create,Le=H.create,Ue=q.create,$e=q.strictCreate,We=Y.create,Ve=X.create,Fe=ee.create,Be=te.create,ze=ne.create,Ke=ae.create,He=re.create,Ge=se.create,qe=ie.create,Ye=oe.create,Je=de.create,Xe=ce.create,Qe=le.create,et=pe.create,tt=he.create,nt=fe.create,at=pe.createWithPreprocess,rt=be.create,st={string:e=>M.create({...e,coerce:!0}),number:e=>j.create({...e,coerce:!0}),boolean:e=>L.create({...e,coerce:!0}),bigint:e=>P.create({...e,coerce:!0}),date:e=>U.create({...e,coerce:!0})},it=f;var ot=Object.freeze({__proto__:null,defaultErrorMap:u,setErrorMap:function(e){d=e},getErrorMap:c,makeIssue:l,EMPTY_PATH:[],addIssueToContext:p,ParseStatus:h,INVALID:f,DIRTY:m,OK:_,isAborted:y,isDirty:v,isValid:g,isAsync:b,get util(){return n},get objectUtil(){return a},ZodParsedType:r,getParsedType:s,ZodType:O,ZodString:M,ZodNumber:j,ZodBigInt:P,ZodBoolean:L,ZodDate:U,ZodSymbol:$,ZodUndefined:W,ZodNull:V,ZodAny:F,ZodUnknown:B,ZodNever:z,ZodVoid:K,ZodArray:H,ZodObject:q,ZodUnion:Y,ZodDiscriminatedUnion:X,ZodIntersection:ee,ZodTuple:te,ZodRecord:ne,ZodMap:ae,ZodSet:re,ZodFunction:se,ZodLazy:ie,ZodLiteral:oe,ZodEnum:de,ZodNativeEnum:ce,ZodPromise:le,ZodEffects:pe,ZodTransformer:pe,ZodOptional:he,ZodNullable:fe,ZodDefault:me,ZodCatch:_e,ZodNaN:ye,BRAND:ve,ZodBranded:ge,ZodPipeline:be,ZodReadonly:xe,custom:ke,Schema:O,ZodSchema:O,late:we,get ZodFirstPartyTypeKind(){return Te},coerce:st,any:Me,array:Le,bigint:Ne,boolean:Ce,date:Ae,discriminatedUnion:Ve,effect:et,enum:Je,function:Ge,instanceof:(e,t={message:`Input not instance of ${e.name}`})=>ke((t=>t instanceof e),t),intersection:Fe,lazy:qe,literal:Ye,map:Ke,nan:Ee,nativeEnum:Xe,never:je,null:Re,nullable:nt,number:Ze,object:Ue,oboolean:()=>Ce().optional(),onumber:()=>Ze().optional(),optional:tt,ostring:()=>Oe().optional(),pipeline:rt,preprocess:at,promise:Qe,record:ze,set:He,strictObject:$e,string:Oe,symbol:Ie,transformer:et,tuple:Be,undefined:Se,union:We,unknown:De,void:Pe,NEVER:it,ZodIssueCode:i,quotelessJson:e=>JSON.stringify(e,null,2).replace(/"([^"]+)":/g,"$1:"),ZodError:o}),ut=Object.defineProperty,dt=(e,t)=>{for(var n in t)ut(e,n,{get:t[n],enumerable:!0})};function ct(e,t=!1){let n=window.crypto.getRandomValues(new Uint32Array(1))[0],a=`_${n}`;return Object.defineProperty(window,a,{value:n=>(t&&Reflect.deleteProperty(window,a),null==e?void 0:e(n)),writable:!1,configurable:!0}),n}async function lt(e,t={}){return new Promise(((n,a)=>{let r=ct((e=>{n(e),Reflect.deleteProperty(window,`_${s}`)}),!0),s=ct((e=>{a(e),Reflect.deleteProperty(window,`_${r}`)}),!0);window.__TAURI_IPC__({cmd:e,callback:r,error:s,...t})}))}function pt(e,t="asset"){let n=encodeURIComponent(e);return navigator.userAgent.includes("Windows")?`https://${t}.localhost/${n}`:`${t}://localhost/${n}`}async function ht(e){return lt("tauri",e)}dt({},{convertFileSrc:()=>pt,invoke:()=>lt,transformCallback:()=>ct});async function ft(e,t){return ht({__tauriModule:"Event",message:{cmd:"unlisten",event:e,eventId:t}})}async function mt(e,t,n){return ht({__tauriModule:"Event",message:{cmd:"listen",event:e,windowLabel:t,handler:ct(n)}}).then((t=>async()=>ft(e,t)))}dt({},{TauriEvent:()=>_t,emit:()=>gt,listen:()=>yt,once:()=>vt});var _t=(e=>(e.WINDOW_RESIZED="tauri://resize",e.WINDOW_MOVED="tauri://move",e.WINDOW_CLOSE_REQUESTED="tauri://close-requested",e.WINDOW_CREATED="tauri://window-created",e.WINDOW_DESTROYED="tauri://destroyed",e.WINDOW_FOCUS="tauri://focus",e.WINDOW_BLUR="tauri://blur",e.WINDOW_SCALE_FACTOR_CHANGED="tauri://scale-change",e.WINDOW_THEME_CHANGED="tauri://theme-changed",e.WINDOW_FILE_DROP="tauri://file-drop",e.WINDOW_FILE_DROP_HOVER="tauri://file-drop-hover",e.WINDOW_FILE_DROP_CANCELLED="tauri://file-drop-cancelled",e.MENU="tauri://menu",e.CHECK_UPDATE="tauri://update",e.UPDATE_AVAILABLE="tauri://update-available",e.INSTALL_UPDATE="tauri://update-install",e.STATUS_UPDATE="tauri://update-status",e.DOWNLOAD_PROGRESS="tauri://update-download-progress",e))(_t||{});async function yt(e,t){return mt(e,null,t)}async function vt(e,t){return async function(e,t,n){return mt(e,t,(t=>{n(t),ft(e,t.id).catch((()=>{}))}))}(e,null,t)}async function gt(e,t){return async function(e,t,n){await ht({__tauriModule:"Event",message:{cmd:"emit",event:e,windowLabel:t,payload:n}})}(e,void 0,t)}var bt="plugin:clipboard|start_monitor",xt="plugin:clipboard|stop_monitor",kt="plugin:clipboard://text-changed",wt="plugin:clipboard://html-changed",Tt="plugin:clipboard://rtf-changed",Ot="plugin:clipboard://files-changed",Zt="plugin:clipboard://image-changed",Et="plugin:clipboard|is_monitor_running",Nt="plugin:clipboard|has_text",Ct="plugin:clipboard|has_image",At="plugin:clipboard|has_html",It="plugin:clipboard|has_rtf",St="plugin:clipboard|write_text",Rt="plugin:clipboard|write_html",Mt="plugin:clipboard|write_rtf",Dt="plugin:clipboard|clear",jt="plugin:clipboard|read_text",Pt="plugin:clipboard|read_html",Lt="plugin:clipboard|read_rtf",Ut="plugin:clipboard|read_files",$t="plugin:clipboard|read_image_binary",Wt="plugin:clipboard|read_image_base64",Vt="plugin:clipboard|write_image_binary",Ft="plugin:clipboard|write_image_base64",Bt="plugin:clipboard://clipboard-monitor/status",zt="plugin:clipboard://clipboard-monitor/update",Kt=ot.object({value:ot.string()}),Ht=ot.object({value:ot.string().array()});function Gt(){return lt(Nt)}function qt(){return lt(At)}function Yt(){return lt(It)}function Jt(){return lt(Ct)}function Xt(e){return lt(St,{text:e})}function Qt(e){return lt(Rt,{html:e})}function en(e){return lt(Mt,{rtf:e})}function tn(){return lt(Dt)}function nn(){return lt(jt)}function an(){return lt(Pt)}function rn(){return lt(Lt)}function sn(){return lt(Ut)}function on(){return lt(Wt)}function un(e){return lt($t).then((function(t){switch(e){case"int_array":default:return t;case"Uint8Array":return new Uint8Array(t);case"Blob":return new Blob([new Uint8Array(t)])}}))}function dn(){return un("Blob").then((function(e){return URL.createObjectURL(e)}))}function cn(e){return lt(Ft,{base64Image:e})}function ln(e){return lt(Vt,{bytes:e})}function pn(n){void 0===n&&(n=500);var a="",r=!0;return setTimeout((function s(){return e(this,void 0,void 0,(function(){var e;return t(this,(function(t){switch(t.label){case 0:return t.trys.push([0,4,,5]),[4,nn()];case 1:return e=t.sent(),a===e?[3,3]:[4,gt(kt,{value:e})];case 2:t.sent(),t.label=3;case 3:return a=e,[3,5];case 4:return t.sent(),[3,5];case 5:return r&&setTimeout(s,n),[2]}}))}))}),n),function(){r=!1}}function hn(n){void 0===n&&(n=1e3);var a="",r=!0;return setTimeout((function s(){return e(this,void 0,void 0,(function(){var e;return t(this,(function(t){switch(t.label){case 0:return t.trys.push([0,4,,5]),[4,on()];case 1:return e=t.sent(),a===e?[3,3]:[4,gt(Zt,{value:e})];case 2:t.sent(),t.label=3;case 3:return a=e,[3,5];case 4:return t.sent(),[3,5];case 5:return r&&setTimeout(s,n),[2]}}))}))}),n),function(){r=!1}}function fn(){var n=this;return yt(zt,(function(a){return e(n,void 0,void 0,(function(){var e,n,r,s,i,o,u,d,c,l,p,h;return t(this,(function(t){switch(t.label){case 0:if("clipboard update"!==a.payload)return[3,22];t.label=1;case 1:return t.trys.push([1,4,,22]),[4,sn()];case 2:return e=t.sent(),[4,gt(Ot,{value:e})];case 3:return t.sent(),[3,22];case 4:return t.sent(),n=!1,[4,Jt()];case 5:return t.sent()?[4,on()]:[3,9];case 6:return(r=t.sent())?[4,gt(Zt,{value:r})]:[3,8];case 7:t.sent(),t.label=8;case 8:n=!0,t.label=9;case 9:return[4,qt()];case 10:return t.sent()?(s=gt,i=[wt],l={},[4,an()]):[3,13];case 11:return[4,s.apply(void 0,i.concat([(l.value=t.sent(),l)]))];case 12:t.sent(),n=!0,t.label=13;case 13:return[4,Yt()];case 14:return t.sent()?(o=gt,u=[Tt],p={},[4,rn()]):[3,17];case 15:return[4,o.apply(void 0,u.concat([(p.value=t.sent(),p)]))];case 16:t.sent(),n=!0,t.label=17;case 17:return[4,Gt()];case 18:return t.sent()?(d=gt,c=[kt],h={},[4,nn()]):[3,21];case 19:return[4,d.apply(void 0,c.concat([(h.value=t.sent(),h)]))];case 20:t.sent(),n=!0,t.label=21;case 21:if(!n)throw new Error("Unexpected Error: No proper clipboard type");return[3,22];case 22:return[2]}}))}))}))}function mn(e){return yt(zt,e)}function _n(n){return e(this,void 0,void 0,(function(){return t(this,(function(e){switch(e.label){case 0:return[4,yt(kt,(function(e){var t=Kt.parse(e.payload).value;n(t)}))];case 1:return[2,e.sent()]}}))}))}function yn(n){return e(this,void 0,void 0,(function(){return t(this,(function(e){switch(e.label){case 0:return[4,yt(wt,(function(e){var t=Kt.parse(e.payload).value;n(t)}))];case 1:return[2,e.sent()]}}))}))}function vn(n){return e(this,void 0,void 0,(function(){return t(this,(function(e){switch(e.label){case 0:return[4,yt(Tt,(function(e){var t=Kt.parse(e.payload).value;n(t)}))];case 1:return[2,e.sent()]}}))}))}function gn(n){return e(this,void 0,void 0,(function(){return t(this,(function(e){switch(e.label){case 0:return[4,yt(Ot,(function(e){var t=Ht.parse(e.payload).value;n(t)}))];case 1:return[2,e.sent()]}}))}))}function bn(n){return e(this,void 0,void 0,(function(){return t(this,(function(e){switch(e.label){case 0:return[4,yt(Zt,(function(e){var t=Kt.parse(e.payload).value;n(t)}))];case 1:return[2,e.sent()]}}))}))}function xn(){return lt(Et).then((function(e){return ot.boolean().parse(e)}))}function kn(){return lt(bt)}function wn(){return lt(xt)}function Tn(n){return e(this,void 0,void 0,(function(){return t(this,(function(e){switch(e.label){case 0:return[4,yt(Bt,(function(e){var t=ot.boolean().parse(e.payload);n(t)}))];case 1:return[2,e.sent()]}}))}))}function On(){var n=this;return kn().then((function(){return fn()})).then((function(a){return function(){return e(n,void 0,void 0,(function(){return t(this,(function(e){switch(e.label){case 0:return a(),[4,wn()];case 1:return e.sent(),[2]}}))}))}}))}var Zn=Object.freeze({__proto__:null,START_MONITOR_COMMAND:bt,STOP_MONITOR_COMMAND:xt,TEXT_CHANGED:kt,HTML_CHANGED:wt,RTF_CHANGED:Tt,FILES_CHANGED:Ot,IMAGE_CHANGED:Zt,IS_MONITOR_RUNNING_COMMAND:Et,HAS_TEXT_COMMAND:Nt,HAS_IMAGE_COMMAND:Ct,HAS_HTML_COMMAND:At,HAS_RTF_COMMAND:It,WRITE_TEXT_COMMAND:St,WRITE_HTML_COMMAND:Rt,WRITE_RTF_COMMAND:Mt,CLEAR_COMMAND:Dt,READ_TEXT_COMMAND:jt,READ_HTML_COMMAND:Pt,READ_RTF_COMMAND:Lt,READ_FILES_COMMAND:Ut,READ_IMAGE_BINARY_COMMAND:$t,READ_IMAGE_BASE64_COMMAND:Wt,WRITE_IMAGE_BINARY_COMMAND:Vt,WRITE_IMAGE_BASE64_COMMAND:Ft,CLIPBOARD_MONITOR_STATUS_UPDATE_EVENT:Bt,MONITOR_UPDATE_EVENT:zt,ClipboardChangedPayloadSchema:Kt,ClipboardChangedFilesPayloadSchema:Ht,hasText:Gt,hasHTML:qt,hasRTF:Yt,hasImage:Jt,writeText:Xt,writeHtml:Qt,writeRtf:en,clear:tn,readText:nn,readHtml:an,readRtf:rn,readFiles:sn,readImageBase64:on,readImageBinary:un,readImageObjectURL:dn,writeImageBase64:cn,writeImageBinary:ln,startBruteForceTextMonitor:pn,startBruteForceImageMonitor:hn,listenToClipboard:fn,onClipboardUpdate:mn,onTextUpdate:_n,onHTMLUpdate:yn,onRTFUpdate:vn,onFilesUpdate:gn,onImageUpdate:bn,isMonitorRunning:xn,startMonitor:kn,stopMonitor:wn,listenToMonitorStatusUpdate:Tn,startListening:On});export{Dt as CLEAR_COMMAND,Bt as CLIPBOARD_MONITOR_STATUS_UPDATE_EVENT,Ht as ClipboardChangedFilesPayloadSchema,Kt as ClipboardChangedPayloadSchema,Ot as FILES_CHANGED,At as HAS_HTML_COMMAND,Ct as HAS_IMAGE_COMMAND,It as HAS_RTF_COMMAND,Nt as HAS_TEXT_COMMAND,wt as HTML_CHANGED,Zt as IMAGE_CHANGED,Et as IS_MONITOR_RUNNING_COMMAND,zt as MONITOR_UPDATE_EVENT,Ut as READ_FILES_COMMAND,Pt as READ_HTML_COMMAND,Wt as READ_IMAGE_BASE64_COMMAND,$t as READ_IMAGE_BINARY_COMMAND,Lt as READ_RTF_COMMAND,jt as READ_TEXT_COMMAND,Tt as RTF_CHANGED,bt as START_MONITOR_COMMAND,xt as STOP_MONITOR_COMMAND,kt as TEXT_CHANGED,Rt as WRITE_HTML_COMMAND,Ft as WRITE_IMAGE_BASE64_COMMAND,Vt as WRITE_IMAGE_BINARY_COMMAND,Mt as WRITE_RTF_COMMAND,St as WRITE_TEXT_COMMAND,tn as clear,Zn as default,qt as hasHTML,Jt as hasImage,Yt as hasRTF,Gt as hasText,xn as isMonitorRunning,fn as listenToClipboard,Tn as listenToMonitorStatusUpdate,mn as onClipboardUpdate,gn as onFilesUpdate,yn as onHTMLUpdate,bn as onImageUpdate,vn as onRTFUpdate,_n as onTextUpdate,sn as readFiles,an as readHtml,on as readImageBase64,un as readImageBinary,dn as readImageObjectURL,rn as readRtf,nn as readText,hn as startBruteForceImageMonitor,pn as startBruteForceTextMonitor,On as startListening,kn as startMonitor,wn as stopMonitor,Qt as writeHtml,cn as writeImageBase64,ln as writeImageBinary,en as writeRtf,Xt as writeText};
